Custom operation

Custom operation on a coin. This is an experimental new feature so please check the results carefully.

Usage

# S3 method for coin
Custom(
  x,
  dset,
  f_cust,
  f_cust_para = NULL,
  write_to = NULL,
  write2log = TRUE,
  ...
)

Arguments

x

A coin

dset

Target data set

f_cust

Function to apply to the data set. See details.

f_cust_para

Optional additional parameters to pass to the function defined by f_cust.

write_to

Name of data set to write to

write2log

Logical: whether or not to write to the log.

...

Arguments to pass to/from other methods.

Value

A coin

Details

In this function, the data set named dset is extracted from the coin using get_dset(coin, dset). It is passed to the function f_cust, which is required to return an equivalent but modified data frame, which is then written as a new data set with name write_to. This is intended to allow arbitrary operations on coin data sets while staying within the COINr framework, which means that if Regen() is used, these operations will be re-run, allowing them to be included in things like sensitivity analysis.

The format of f_cust is important. It must be a function whose first argument is called x: this will be the argument that the data is passed to. The data will be in the same format as extracted via get_dset(coin, dset), which means it will have a uCode column. f_cust can have other arguments which are passed to it via f_cust_para. The function should return a data frame similar to the data that was passed to it, it must contain have the same column names (meaning you can't remove indicators), but otherwise is flexible - this means some caution is necessary to ensure that subsequent operations don't fail. Be careful, for example, to ensure that there are no duplicates in uCode, and that indicator columns are numeric.

Examples

# build example coin
coin <- build_example_coin(up_to = "new_coin")
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

# create function - replaces suspected unreliable point with NA
f_NA <- function(x){ x[3, 10] <- NA; return(x)}

# call function from Custom()
coin <- Custom(coin, dset = "Raw", f_cust = "f_NA")
#> Error in f_NA(x = structure(list(uCode = c("AUS", "AUT", "BEL", "BGD", "BGR", "BRN", "CHE", "CHN", "CYP", "CZE", "DEU", "DNK", "ESP", "EST", "FIN", "FRA", "GBR", "GRC", "HRV", "HUN", "IDN", "IND", "IRL", "ITA", "JPN", "KAZ", "KHM", "KOR", "LAO", "LTU", "LUX", "LVA", "MLT", "MMR", "MNG", "MYS", "NLD", "NOR", "NZL", "PAK", "PHL", "POL", "PRT", "ROU", "RUS", "SGP", "SVK", "SVN", "SWE", "THA", "VNM"), LPI = c(3.793385, 4.097985, 4.108538, 2.663902, 2.807685, 2.870492, 3.987158, 3.661104, 2.999061, 3.674309, 4.225967, 3.815794, 3.727412, 3.363489, 3.920745, 3.900953, 4.069669, 3.239516, 3.160829, 3.428968, 2.984537, 3.420043, 3.794886, 3.755414, 3.970464, 2.751998, 2.80059, 3.717126, 2.067254, 3.631688, 4.219409, 3.327107, 3.069256, 2.458571, 2.506056, 3.426307, 4.18753, 3.732163, 3.388, 2.923219, 2.856259, 3.425877, 3.409367, 2.99312, 2.570864, 4.143632, 3.336895, 3.184508, 4.204593, 3.2551, 2.976629), Flights = c(36.05498, 29.01725, 31.88546, 4.27955, 9.23588, 2.019, 51.78846, 114.2008, 8.75467, 15.30953, 174.3588, 32.7705, 170.9628, 3.12946, 18.85806, 97.63489, 210.8244, 34.83849, 9.24529, 13.01196, 34.6802, 25.2643, 34.17721, 110.2755, 69.00941, 4.02303, 9.5212, 69.84827, 3.0744, 5.37919, 4.84458, 6.77976, 6.75251, 6.695, 0.98951, 53.33988, 63.59241, 25.64994, 13.37242, 2.21146, 19.43838, 33.79735, 40.46484, 18.79894, 34.06447, 67.86156, 2.07397, 1.51736, 31.74444, 73.92221, 28.52418), Ship = c(14.0041976, 0, 20.5671215, 9.698165, 7.9193664, 7.4925931, 0, 21.1719761, 11.6894955, 0, 20.5012617, 14.4070803, 19.3288348, 9.1719971, 10.0441551, 19.5361146, 20.8004246, 15.6795083, 12.4404524, 0, 12.2099979, 16.5221178, 10.5751931, 18.3837783, 17.2613375, 0, 9.0174744, 20.4274184, 0, 9.2343492, 0, 7.8509367, 17.0605518, 9.1824031, 0, 20.551337, 20.5415482, 9.2878889, 11.9107463, 15.3199989, 11.6212278, 14.8155168, 16.1988714, 12.6898376, 15.8114854, 20.2768725, 0, 13.3801771, 15.5780928, 14.5063989, 16.5487609), Bord = c(0, 35, 48, 16, 18, 2, 41, 32, 0, 35, 105, 3, 26, 6, 18, 75, 13, 12, 41, 28, 2, 35, 12, 24, 0, 46, 13, 1, 20, 22, 12, 17, 0, 8, 9, 9, 32, 19, 0, 43, 0, 62, 11, 24, 122, 2, 25, 20, 18, 17, 25), Elec = c(0, 35.3697298333333, 26.5330466666667, 1.33161666666667, 11.2775841666667, 0, 68.7732871666667, 6.43630583333333, 0.439364333333333, 46.7581786666667, 110.179562833333, 20.103365, 24.1796233333333, 9.49907183333333, 24.4142605, 78.5259303333333, 25.7837076666667, 2.43086016666667, 19.528362, 20.5432411666667, 2e-06, 1.33225166666667, 3.39668933333333, 49.3295443333333, 3.6e-05, 3.08950333333333, 2.1845235, 3.6e-05, 2.047842, 7.4940275, 6.91123783333333, 5.681299, 1.06719433333333, 0.3967805, 1.52714166666667, 0.0673386666666667, 44.0901205, 19.5413745, 0, 0.003016, 0, 26.3441995, 10.1248156666667, 8.264701, 14.188028, 2e-06, 18.8196528333333, 21.6560726666667, 42.4725213333333, 2.25995316666667, 3.524199), Gas = c(1.14, 0.273, 36.1, 0.144, 0.312, 6.04, 3.05, 9.46, 0.029, 2.7, 48.7, 0.418, 3.29, 0.0667, 0.38, 29.9, 42.5, 0.139, 0.422, 0.419, 22.3, 0.751, 2.17, 4.22, 34.5, 12.7, 0.135, 12.5, 0.00841, 0.98, 0.482, 0.377, 0.0719, 24.1, 0.0267, 26.3, 23, 94.8, 0.0226, 0.195, 0.683, 3.68, 0.573, 0.376, 19.8, 9.43, 4.63, 0.442, 1.69, 24.5, 0.776), ConSpeed = c(11.1, 14.1, 16.3, NA, 15.5, NA, 21.7, 7.6, 6.9, 16.9, 15.3, 20.1, 15.5, 11.6, 20.5, 10.8, 16.9, 7.9, 8.6, 14.8, 7.2, 6.5, 15.6, 9.2, 20.2, NA, NA, 28.6, NA, 14.6, 11.6, 16.6, 12.8, NA, NA, 8.9, 17.4, 23.5, 14.7, NA, 5.5, 12.6, 12.9, 17, 11.8, 20.3, 13, 13.7, 22.5, 16, 9.5),     Cov4G = c(94, 98, 99.89, 65, 56.73, 80, 98, 85, 60, 98.7,     95.7, 99.99, 91.3, 100, 99.9, 80, 97.8, 83, 98, 97.3, 5,     4, 90, 93, 99, 65.5, 30, 99, 5, 91, 96, 90, 100, 0, 90, 71,     99, 99, 88, 16, 39, 100, 94.3, 72, 50, 100, 75, 97.7, 99.99,     21, 0), Goods = c(288.4893, 278.4264, 597.8723, 56.24166,     42.82515, 7.23391, 402.8649, 1713.619, 8.76681, 274.1365,     1919.194, 146.6771, 447.1229, 28.2411, 101.575, 849.3303,     778.9052, 56.62702, 28.36795, 173.6159, 222.4186, 288.9806,     139.9937, 658.1981, 732.2078, 55.04856, 20.75014, 568.992,     8.95527, 42.12889, 30.97279, 28.29194, 13.2783, 31.70398,     7.81803, 278.0319, 770.3935, 142.6095, 48.20846, 45.28752,     118.5618, 363.0132, 102.4915, 115.3185, 343.8504, 414.2389,     140.366, 54.98977, 233.3067, 282.8014, 269.0766), Services = c(108.6645,     108.1273, 216.3146, 10.07891, 12.97674, 2.11464, 207.3637,     657.109, 15.23712, 43.46382, 578.4264, 113.6753, 197.1981,     10.2278, 53.80751, 471.2931, 518.2256, 38.17887, 17.35676,     39.11952, 54.06682, 294.2822, 338.0985, 202.5398, 351.4253,     16.85995, 5.79267, 200.8481, 1.37855, 12.35732, 165.791,     7.25574, 20.06847, 6.24003, 2.72198, 72.93042, 346.5511,     83.35802, 26.55097, 11.00249, 55.21572, 82.74544, 43.77949,     31.05017, 122.5508, 304.7169, 16.2451, 11.79091, 132.1061,     108.0109, 30.53244), FDI = c(20.9, 5, 5.71, 3.01, 1.35, 0.13,     12.2, 75.6, 1.23, 3.88, 47, 9.1, 17.1, 0.58, 6.03, 30.9,     55.5, 0.883, 0.387, 3.21, 19.3, 41.8, 6.1, 10.3, 37.1, 4.99,     4.19, 26.7, 1.75, 0.965, 7.99, 0.444, 0.179, 6.7, 2.1, 21.9,     13.9, 4.57, 1.85, 6.95, 6.13, 10.1, 2.33, 4.05, 17, 27, 2.32,     0.362, 8.07, 17.2, 24.9), PRemit = c(14.7103856460272, 5.06901177701532,     13.354527048996, 7.46938906780951, 1.03954138466167, 0.76046276837045,     9.48934118719299, 26.5510246778493, 0.477270474235235, 4.69155756880396,     28.6815952667245, 2.1877554092349, 16.7522764146496, 0.588833699996422,     1.51021445339271, 30.2099053373786, 20.6965328785305, 0.763007509197043,     1.56241188807008, 4.16558155904503, 4.4459418154073, 14.6034747360715,     1.96058590349577, 16.2027497039317, 10.5489455525299, 2.55539529246998,     0.628865667275296, 8.71109216763211, 0.16621267270068, 1.14676288611879,     2.89930363704331, 1.03747608199413, 0.233004395996157, 0.906530108847677,     0.311532362294345, 7.38969263584757, 4.89495380227947, 2.28520295973178,     2.3361618328008, 3.96382219102169, 8.26935692187302, 6.29458465199366,     5.23096443207958, 4.73419428733079, 5.56102499851568, 6.05749642534467,     2.76756018735459, 0.775929007060537, 4.74525818658639, 7.31779689453488,     5.08498071988885), ForPort = c(2188.311, 808.6832, 1574.308,     8.48896, 15.5088, NA, 2615.089, 1541.035, 40.76228, 107.7103,     6386.918, 1021.278, 2040.571, 17.41899, 747.6268, 6745.328,     8206.978, 185.1048, 16.88908, 72.5257, 287.6153, 270.2824,     6908.412, 3264.087, 7449.993, 87.49439, NA, 1195.635, NA,     31.03612, 10601.75, 26.82834, 139.8678, 0.04805, 5.20507,     248.1615, 4697.464, 1632.781, 235.9858, 13.38957, 103.6887,     218.883, 327.1822, 35.2075, 304.7017, 1487.02, 70.29841,     49.05213, 1463.559, 220.1121, 0.00189), Embs = c(82, 88,     84, 52, 67, 41, 82, 100, 43, 84, 100, 77, 89, 46, 74, 100,     100, 83, 62, 79, 83, 95, 74, 93, 95, 62, 42, 86, 41, 50,     37, 48, 28, 52, 40, 74, 89, 78, 44, 72, 66, 87, 72, 75, 99,     58, 62, 43, 78, 78, 80), IGOs = c(196, 227, 248, 145, 209,     85, 222, 193, 172, 201, 309, 259, 280, 194, 269, 329, 285,     216, 197, 212, 175, 198, 195, 288, 201, 117, 95, 210, 82,     190, 190, 184, 169, 87, 107, 177, 265, 259, 149, 157, 158,     238, 236, 217, 232, 119, 201, 197, 278, 152, 138), UNVote = c(38.4624535523426,     42.6391988182863, 43.0030841510482, 38.6060071756095, 42.9598569072775,     38.5010926030476, 42.6412702963947, 37.1366126202482, 42.3470518317134,     42.2250633366699, 42.6261490865325, 42.7914044892956, 43.1010323381268,     42.8799747579199, 42.7462886045531, 40.4229453212477, 39.4011794236269,     42.6930073737805, 42.7496555271324, 42.5335083569395, 38.1047713110564,     36.9213095801954, 42.6856751008368, 42.9949093382342, 42.1723082984885,     38.7268707722602, 38.0223159025561, 42.0001157307494, 37.6329113519425,     42.7128277596374, 42.984261891972, 42.7006553552221, 42.5648464970049,     35.7557197463119, 39.052573350744, 38.2849251695488, 42.8032134532135,     43.1111390764364, 42.3222171664533, 38.1493888682547, 37.6816035761387,     42.5849037301029, 43.1860466576031, 42.5248656451369, 36.129704753258,     38.7474998951008, 43.1010323381268, 43.1010323381268, 42.7515289570839,     38.8262888723369, 37.5306259724372), CostImpEx = c(364, 0,     0, 595, 52, 140, 150, 255.5, 100, 0, 45, 0, 0, 0, 70, 0,     25, 30, 0, 0, 303.2, 226.7, 150, 0, 161, 320, 220, 38, 350,     28, 0, 35, 25, 350, 147, 105, 0, 0, 147, 992, 103, 0, 0,     0, 244.5, 77, 0, 0, 40, 140, 322), Tariff = c(1.17, 1.6,     1.6, 10.53, 1.6, 0.5, 0, 3.54, 1.6, 1.6, 1.6, 1.6, 1.6, 1.6,     1.6, 1.6, 1.6, 1.6, 1.6, 1.6, 2.64, 6.32, 1.6, 1.6, 1.35,     2.51, 9.77, 7.68, 1.65, 1.6, 1.6, 1.6, 1.6, NA, NA, 3.68,     1.6, 1.02, 1.27, 9.99, 3.4, 1.6, 1.6, 1.6, 3.43, 0, 1.6,     1.6, 1.6, NA, 2.86), TBTs = c(205, 1144, 1348, NA, 1140,     2, 300, 1230, 1141, 1456, 1163, 1393, 1212, 1153, 1215, 1385,     1187, 1140, 1179, 1173, 118, 117, 1143, 1171, 805, 19, 3,     839, 1, 1170, 1140, 1170, 1140, 2, 7, 236, 1754, 81, 108,     112, 256, 1147, 1141, 1230, 85, 57, 1187, 1250, 1364, 609,     114), TIRcon = c(0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1,     0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0), RTAs = c(14,     30, 30, 12, 30, 15, 33, 17, 30, 30, 30, 30, 30, 30, 30, 30,     30, 30, 30, 30, 17, 15, 30, 30, 14, 2, 15, 46, 16, 30, 30,     30, 30, 17, 1, 17, 30, 31, 13, 11, 17, 30, 30, 30, 2, 19,     30, 30, 30, 17, 19), Visa = c(68, 80, 80, 45, 79, 90, 80,     11, 79, 79, 82, 81, 81, 79, 81, 81, 81, 80, 78, 79, 59, 51,     80, 81, 84, 48, 61, 85, 61, 79, 80, 79, 80, 17, 21, 91, 80,     81, 80, 1, 57, 79, 80, 79, 17, 92, 79, 79, 81, 61, 31), StMob = c(268,     69.8, 42.3, 50.4, 28.8, 3.5, 49.8, 445, 29.4, 47.9, 233,     33.2, 51, 5.99, 22, 152, 326, 48.6, 5.05, 26.2, 31.1, 120,     24.5, 97, 133, 83.9, 4.43, 83.8, 6.46, 12.1, 12.7, 8.79,     1.77, 5.81, 8.48, 115, 69.3, 22.6, 49.2, 30.3, 10.6, 35.6,     14.4, 38.7, 135, 19.3, 40.9, 3.6, 28.1, 30.3, 53.4), Research = c(46761,     14432, 20176, 2084, 1854, 229, 30056, 95919, 1337, 9190,     78353, 15986, 39777, 1594, 10303, 56677, 96337, 8020, 2342,     5180, 2624, 20234, 7451, 45825, 31184, 1026, 378, 20437,     175, 1482, 1206, 794, 347, 299, 293, 8080, 33445, 11696,     7731, 7122, 1361, 13008, 10943, 4173, 16182, 11411, 2741,     2653, 23514, 5317, 3618), Pat = c(229.8, 519.1, 655.8, NA,     12.2, NA, 1315.7, 927.8, 7.4, 112.6, 2771.7, 306, 334.7,     14.4, 412.5, 1469, 1313.7, 43.4, 15.1, 109.5, 23.7, 437.6,     140.6, 513, 583.5, 3.6, NA, 249.8, NA, 7.5, 66.2, 1.5, 4.7,     NA, 0.3, 64.2, 641.2, 168.1, 46.5, 7.2, 11.3, 179.3, 46.1,     42.6, 141.5, 270.5, 44.2, 30.8, 661.5, 53.6, NA), CultServ = c(2.00098,     1.52758, 2.38787, 0.01472, 0.11991, NA, 1.15976, 2.88624,     0.137, 0.3223, 4.83603, 1.98816, NA, 0.10275, 0.58878, 7.01905,     9.56996, 0.47483, 0.28331, 1.51071, 0.13405, 3.29198, 0.5222,     0.57832, 2.20754, 0.0578, 0.00501, 1.798, NA, 0.06532, 8.045,     0.0332, 2.46272, 0.08905, 0.00266, 1.15292, 2.42317, 1.66163,     0.34615, 0.03553, 0.29555, 1.44714, 0.53946, 0.1449, 1.44633,     0.9278, 0.0882, 0.20211, 0.90896, 0.08969, NA), CultGood = c(6.807,     8.932, 12.698, NA, 0.88, 0.364, 28.965, 74.468, 0.529, 8.807,     47.161, 4.912, 11.759, 0.745, 1.858, 41.028, 35.139, 1.626,     0.737, 2.614, NA, 7.991, 3.424, 26.843, 16.182, 0.985, NA,     NA, NA, 1.281, 0.636, 0.735, 0.369, NA, 0.046, 7.555, 19.159,     2.981, 1.213, 1.256, 3.185, 11.912, 2.616, 2.387, 8.379,     14.507, 2.407, 1.212, 6.014, 6.661, NA), Tourist = c(8.263,     28.121, 7.481, 0.125, 8.252, 0.219, 9.205, 59.27, 3.187,     9.321, 35.555, 10.781, 75.315, 3.147, 2.789, 82.57, 35.814,     24.799, 13.809, 5.302, 11.519, 14.569, 10.1, 52.372, 24.04,     6.509, 5.012, 17.242, 3.315, 2.296, 1.054, 1.793, 1.966,     2.907, 0.404, 26.757, 15.828, 5.96, 3.37, 0.966, 5.967, 17.471,     11.223, 10.223, 24.571, 12.914, 2.027, 3.032, 6.782, 32.53,     10.013), MigStock = c(5.66849, 1.30365, 1.32919, 5.08139,     0.69462, 0.14041, 2.12634, 4.30982, 0.23838, 1.10637, 10.9285,     0.54405, 3.34468, 0.32117, 0.46894, 4.37339, 8.91445, 1.06705,     0.54069, 0.84562, 2.06821, 8.43759, 1.28606, 4.87039, 2.19471,     6.38381, 0.93047, 2.06535, 1.12532, 0.5749, 0.29738, 0.47182,     0.11826, 2.57501, 0.0817, 4.00744, 1.68457, 0.66093, 1.48606,     4.09286, 1.40403, 4.24078, 1.75075, 3.32536, 7.65393, 2.47645,     0.49825, 0.18037, 1.03933, 4.13663, 1.16007), Lang = c(19.0054374598599,     14.1904159671892, 10.690851815187, 0.177440480421409, 6.28997760955522,     6.24168793798948, 13.6355487327224, 0.73963573832551, 12.3211798230603,     7.72876772483642, 15.2613495482733, 19.2699693892385, 5.15846746249101,     12.2115279608916, 15.0435333505633, 11.5031818152116, 19.352130196013,     10.5901398719892, 11.9764149171441, 5.22493581171035, 0.0322205918335903,     3.77194294687334, 19.4642552966452, 6.40814648739812, 0.0310785114043021,     7.20274576913582, 4.4104924825484, 2.1346106433082, 0.166383539236451,     10.5004232127186, 21.4852605620001, 12.0078424736489, 16.9119582606504,     0.0188484825710535, 2.25185464024358, 6.66895072968037, 18.1532118955105,     15.0962650463399, 16.6801312858093, 1.71805798047815, 10.8942511388064,     9.10344745704517, 6.93615852372676, 7.05215978791605, 5.92039782459955,     12.5796736508144, 8.23424297957741, 14.1024299989425, 18.167424717557,     0.159059459126493, 0.0850023118396344), Renew = c(9.18050295980786,     34.3949917477854, 9.20164053078148, 34.7470690115416, 17.6500779274042,     0.0149350358914578, 25.2919836825127, 12.4133525698911, 9.9420660163921,     14.8285619496241, 14.2062526977772, 33.1702771730605, 16.254097737239,     27.4768334658828, 43.2352625898503, 13.4992623166677, 8.71168557437913,     17.170047546133, 33.1274666358571, 15.559362050795, 36.8793481032359,     36.0212225652817, 9.08139901337063, 16.5168505774397, 6.29735708059289,     1.55844291225827, 64.9237759106607, 2.70770280422964, 59.3158607506634,     28.9611792543763, 9.03131366096732, 38.0980171666829, 5.35500474641897,     61.5278116473242, 3.42973728977845, 5.19444258901529, 5.88946307882733,     57.7720015985381, 30.7891755544734, 46.4763249199424, 27.4515444500126,     11.9114882087605, 27.15729838594, 23.6977940316245, 3.30422835092973,     0.70863529719762, 13.4091682209694, 20.8773961432658, 53.2477697984376,     22.8630701280756, 34.9985698382085), PrimEner = c(123.010033139841,     84.9278479679633, 113.768713902959, 74.7453254060098, 151.981833383824,     113.449948607868, 53.4785426888034, 175.310890661362, 77.6944953843903,     134.449351949147, 86.7632734881106, 63.7541074347429, 79.0065482864909,     169.383700540124, 159.245257668909, 97.4647297776276, 72.5938294118918,     88.1957676326141, 94.2501021423577, 95.7930413300266, 88.3638018596934,     118.263439272263, 56.0668518295205, 71.1275291972484, 92.9929858856784,     188.009312979212, 133.44976023901, 158.239841759835, NA,     90.9385611437279, 73.2591125591622, 97.9830650051963, 55.9967149255383,     77.9597416149806, 161.978727800047, 122.652870702966, 94.7158162649436,     88.2326566353308, 132.290546561376, 105.861273227578, 72.3128224683811,     101.589493595927, 78.1792293083991, 80.9306398400042, 192.450829415161,     63.7790377253459, 107.867269682028, 113.881172841511, 112.445148432818,     132.604326487521, 130.355426259171), CO2 = c(15.3701377604233,     6.873713220337, 8.32815989783976, 0.459141964802809, 5.87161586934993,     22.1247012416688, 4.31156299409097, 7.54390764144927, 5.26035204098901,     9.16597837581982, 8.88937039483839, 5.9357124821143, 5.03382448723757,     14.848819173174, 8.66072124349448, 4.5720882491356, 6.49744049045117,     6.18033726778447, 3.97380490559031, 4.2655749757664, 1.81936331905513,     1.73000043156161, 7.37811780885705, 5.27086678640296, 9.53870610327163,     14.3623896595937, 0.437759998009271, 11.5703454093401, 0.297200883705774,     4.37809012309851, 17.3621213728095, 3.49819288166911, 5.4915247891727,     0.416600361658081, 7.12732634813277, 8.03299157864044, 9.92013807523839,     9.27094513154166, 7.68657582544293, 0.896264105182138, 1.05545684592961,     7.51715158489872, 4.33155402784831, 3.51615369125659, 11.8575277737052,     10.306331909983, 5.66158132774424, 6.21368684468327, 4.47818217821374,     4.62186003163084, 1.80356574966869), MatCon = c(38.38107313,     15.7636163, 15.88494182, 2.577148198, 18.70468756, 23.33175586,     13.52098315, 23.64706745, 24.42632636, 17.11522483, 15.01561584,     16.65192282, 11.99873543, 32.98061726, 24.60578457, 11.82746131,     8.495755209, 11.62461083, 9.791906479, 16.9231686, 7.190287275,     5.337524694, 13.93800372, 10.90480022, 9.381837526, 28.21245712,     4.882918058, 15.90472487, 10.60669617, 14.80379102, 28.15463343,     16.00262107, 16.26107176, 3.300184316, 33.49389267, 18.85690576,     14.30418582, 21.15019582, 24.86179401, 4.432335659, 3.97295926,     17.68296232, 11.0908576, 11.46947366, 16.36200395, 33.36901263,     10.91072076, 13.77068583, 17.02767911, 12.30469194, 13.59489345    ), Forest = c(6.32952941468394, 3.03828098335719, 5.70982767587253,     5.2605369742498, 1.54839009809886, 3.22697538934652, 0.991674932493423,     3.72514537262857, 1.76620676551156, 4.98479832504303, 2.37034887142205,     5.30439350790529, 6.92659910049046, 9.40621321932958, 8.5039376880111,     3.78798081995652, 5.46621928609287, 1.89976598579562, 1.72809763707082,     5.84718093038222, 10.6347953960583, 3.35525381682685, 3.56089418132601,     2.05435874242636, 1.4232737467321, 0.480100582615046, 22.8067645922082,     3.03831722282878, 13.1990071747579, 8.09749723328876, 4.54462959578252,     10.3812688543045, 2.92633199691794, 6.36026041823957, 2.94196578582509,     17.2735412850144, 1.82845479822948, 3.11755115062844, 6.06949099692683,     0.326390728585887, 4.47376862568818, 5.80584309676141, 31.8455434477583,     2.43481051117764, 5.30075011325462, 5.98369061009474, 5.02494037689629,     2.05971130439811, 8.50160666939715, 6.76115648981457, 12.7836337026203    ), Poverty = c(0.3, 0.7, 0, 19.6, 1.5, NA, 0, 1.4, 0, 0,     0, 0.2, 1, 0.5, 0, 0, 0.2, 1.5, 0.7, 0.5, 7.2, 21.2, 0.5,     NA, NA, 0.1, NA, 0.3, 22.7, 0.7, 0.2, 0.7, 0, 6.4, 0.2, NA,     0, 0.2, NA, 6.1, 8.3, 0, 0.5, 0, 0, NA, 0.7, 0, 0.5, 0, 2.6    ), Palma = c(1.4, 1.11, 0.96, 1.26, 1.49, NA, 1.2, 2.1, 1.4,     0.92, 1.1, 1.04, 1.46, 1.26, 0.95, 1.31, 1.25, 1.55, 1.22,     1.13, 1.81, 1.48, 1.25, 1.42, 1.22, 0.92, 1.16, NA, 1.68,     1.43, 1.4, 1.44, NA, NA, 1.23, 2.62, 0.99, 0.88, NA, 1.18,     2.17, 1.23, 1.5, 0.96, 2.02, NA, 0.88, 0.88, 0.93, 1.65,     1.64), TertGrad = c(29.78425, 13.08167, 31.70948, NA, 24.47098,     NA, 36.94929, 3.57765, 25.54672, 19.3418, 24.74205, 30.38373,     29.29108, 37.60498, 22.50554, 17.66307, 29.33262, 22.92907,     18.25773, 20.67309, 8.48301, 9.13629, 26.8378, 14.35266,     29.87681, NA, NA, 24.35774, NA, 33.09735, 34.27727, 27.79356,     13.28486, NA, 23.71332, 16.36954, 28.86907, 26.65009, 26.00903,     8.75331, 16.99506, 23.7002, 18.18253, 13.01408, 2.00915,     28.21053, 18.92215, 18.82353, 22.99885, 13.12545, NA), FreePress = c(22,     22, 12, 62, 42, 76, 13, 87, 23, 21, 20, 12, 28, 16, 12, 26,     25, 44, 41, 44, 49, 43, 18, 31, 27, 85, 70, 34, 85, 21, 14,     26, 23, 73, 37, 69, 11, 8, 19, 65, 44, 34, 17, 38, 83, 67,     26, 23, 11, 77, 84), TolMin = c(3.6, 4.8, 4.6, 8.4, 5.1,     5.6, 3.6, 7.6, 6, 5.1, 4.9, 4.4, 5.8, 7.5363161847766, 1.5,     7, 6.4, 5.1, 5.5, 4.5, 7.2, 8.3, 1.1, 4.8, 3.4, 7.9, 6.6,     2.6, 6.40654892454685, 4.1, 2.8, 8.3, 3.6, 9.8, 3.5, 6.2,     4.5, 3.4, 3.3, 9.7, 8.2, 5.9, 2.2, 6.5, 8.5, 2.3, 6.6, 4.5,     1.8, 8.2, 5.7), NGOs = c(311, 301, 1824, 16, 28, 1, 741,     117, 19, 55, 986, 208, 341, 21, 118, 1071, 1582, 102, 30,     62, 32, 167, 64, 498, 252, 5, 0, 145, 1, 16, 66, 18, 14,     1, 3, 113, 731, 160, 54, 25, 104, 50, 81, 30, 82, 139, 24,     31, 299, 98, 8), CPI = c(77, 75, 75, 28, 43, 62, 85, 41,     57, 57, 81, 88, 57, 71, 85, 70, 82, 48, 49, 45, 37, 40, 74,     50, 73, 31, 21, 54, 29, 59, 82, 58, 56, 30, 36, 47, 82, 85,     89, 32, 34, 60, 63, 48, 29, 84, 50, 61, 84, 37, 35), FemLab = c(0.859266059515803,     0.887213779453269, 0.869972390202379, 0.54140288967241, 0.892763098960666,     NA, 0.893641806962591, 0.835955527204624, 0.863333365381842,     0.811948281457023, 0.88449097961006, 0.926999958117785, 0.863148694124589,     0.900075149299222, 0.965190995604275, 0.895348576860072,     0.869741934200415, 0.77849100340878, 0.866495248468469, 0.824900660030214,     0.614982860933004, 0.345576405858748, 0.807835783985394,     0.736446396156309, 0.779610540759765, 0.901036274239526,     0.878812534566152, 0.731127121283903, 1.02747155772406, 0.939814548659182,     0.83410929479832, 0.923410603386617, 0.632396276811038, NA,     0.835193987298336, 0.649986359552759, 0.874858528479807,     0.948004080460814, 0.885417350668179, 0.297008552917755,     0.649032809600658, 0.820179201539319, 0.910405332568602,     0.770354089124809, 0.864915936620173, 0.797709815914294,     0.81326713914964, 0.905521391646056, 0.949085319278949, 0.814882965429736,     0.919440214001039), WomParl = c(28.6667, 30.6011, 38, 20.2857,     19.1667, 9.0909, 32.5, 23.703, 17.8571, 20, 36.9841, 37.4302,     39.1429, 26.7327, 42, 25.8232, 30, 18.3333, 19.8675, 10.0503,     19.8214, 11.8081, 22.1519, 30.9524, 9.2632, 27.1028, 20.3252,     17, 27.5168, 21.2766, 28.3333, 16, 12.5, 10.1617, 17.1053,     10.3604, 38, 39.645, 34.1667, 20.5882, 29.4521, 28.0435,     34.7826, 20.6687, 15.7778, 23.7624, 20, 36.6667, 43.553,     4.8583, 26.7206), PubDebt = c(37.64, 86.19, 106.05, 31.67,     26.3, 2.81, 45.66, 42.92, 108.87, 40.31, 70.99, 45.53, 99.26,     9.74, 62.51, 96.14, 88.98, 176.94, 86.67, 75.33, 27.3, 69.07,     78.68, 132.71, 247.98, 21.89, 32.54, 37.89, 61.86, 42.78,     21.45, 34.94, 63.96, 34.34, NA, 57.39, 65.12, 27.94, 29.95,     63.57, 34.79, 51.28, 128.98, 39.3, 16.42, 104.68, 52.91,     83.15, 43.4, 42.72, 58.25), PrivDebt = c(202.50016466249,     141.28513440112, 225.72426769539, 40.815016861394, 116.98938450041,     NA, 237.7712164601, 193.43419063017, 345.59966891985, 73.213574270152,     105.91991221514, 223.23740707719, 165.7912852854, 126.58230727932,     181.238153059, 188.68200158198, 169.14574410269, 124.65858012962,     130.34435092198, 92.513949619879, 40.838397758038, 55.037166040262,     299.06964573854, 115.32312997003, 158.42064639399, 52.65278610274,     68.158492108364, 193.24989642247, 19.863133529429, 63.302498428946,     421.38527338142, 96.908181303509, 190.06810924364, 21.86631189634,     85.611348481956, 145.18617160268, 231.32965483085, 252.12094321748,     174.0986232151, 19.67392581864, 56.334452213332, 86.22426931394,     184.73319891749, 28.38030228429, 70.670386929365, 165.48938267246,     92.638184979403, 88.1486264817, 207.2226629029, 130.82867907831,     123.81487498901), GDPGrow = c(0.348880361421024, 2.18677558047276,     1.36616463659274, 6.16478747458362, 4.32022260986446, 0.0285102886987971,     -0.0198229475945908, 6.30396712336683, 0, 4.04300124425265,     1.79471020042621, 1.50309096605672, 2.85707386085544, 4.87884627536938,     2.33629470348814, 1.42667763500899, 1.12929130133277, 1.4977922822242,     3.99021292091018, 4.33844640890226, 3.92307260015163, 5.42924311888073,     6.49733803507186, 1.62982756918014, 1.8800197571837, 2.59552271917623,     5.19299957036061, 2.62122194009741, 5.33713472656754, 5.31639214581796,     -0.677398726046491, 5.56194830322714, 4.1499081098989, 5.40145516218355,     4.2235668410402, 4.43836861173791, 2.5447872426889, 0.998313008074007,     0.864451330210272, 3.65519604299547, 5.0597473929578, 4.53454572198346,     2.99637372240358, 7.57933150121373, 0, 3.52678165524152,     3.22730973370055, 4.91413515967231, 0.818382774779309, 3.64110272110705,     5.72587784766472), RDExp = c(2.20207, 3.07181, 2.45668, NA,     0.95657, NA, 2.96635, 2.06558, 0.45585, 1.94865, 2.87749,     3.01387, 1.21961, 1.49502, 2.90474, 2.23135, 1.70304, 0.95673,     0.85439, 1.37762, 0.08466, 0.6274, 1.51368, 1.33488, 3.28363,     0.16944, 0.11827, 4.22816, NA, 1.04246, 1.28773, 0.6255,     0.76923, NA, 0.15501, 1.29813, 2.01329, 1.93301, 1.15236,     0.24554, 0.13839, 1.00337, 1.27885, 0.48765, 1.13202, 2.19754,     1.17845, 2.21174, 3.26285, 0.6272, 0.37404), NEET = c(8.7,     6.5, 9.3, 27.4, 15.3, 17.2, 6.5, NA, 16.1, 6.3, 6.3, 7, 13.3,     9.4, 9.4, 11.5, 10.3, 15.3, 15.4, 11, 21.5, 27.9, 10.9, 20,     3.5, 9.5, 12.7, NA, 42.1, 9.2, 5.9, 10.3, 8, 17.4, 19.8,     12.8, 4, 4.6, 11.8, 30.4, 21.7, 9.5, 9.3, 15.2, 12.4, 4,     12.1, 6.5, 6.2, 15, 0.6)), row.names = c(31L, 1L, 2L, 32L, 3L, 33L, 29L, 35L, 5L, 6L, 11L, 7L, 27L, 8L, 9L, 10L, 30L, 12L, 4L, 13L, 37L, 36L, 14L, 15L, 38L, 39L, 34L, 40L, 41L, 17L, 18L, 16L, 19L, 44L, 43L, 42L, 20L, 21L, 45L, 46L, 47L, 22L, 23L, 24L, 48L, 49L, 25L, 26L, 28L, 50L, 51L), class = "data.frame")): could not find function "f_NA"
stopifnot(is.na(coin$Data$Custom[3,10]))