Impute a data set in a coin

This imputes any NAs in the data set specified by dset by invoking the function f_i and any optional arguments f_i_para on each column at a time (if impute_by = "column"), or on each row at a time (if impute_by = "row"), or by passing the entire data frame to f_i if impute_by = "df".

Usage

# S3 method for coin
Impute(
  x,
  dset,
  f_i = NULL,
  f_i_para = NULL,
  impute_by = "column",
  use_group = NULL,
  group_level = NULL,
  normalise_first = NULL,
  out2 = "coin",
  write_to = NULL,
  ...
)

Arguments

x

A coin class object

dset

The name of the data set to apply the function to, which should be accessible in .$Data.

f_i

An imputation function. See details.

f_i_para

Further arguments to pass to f_i, other than x. See details.

impute_by

Specifies how to impute: if "column", passes each column (indicator) separately as a numerical vector to f_i; if "row", passes each row separately; and if "df" passes the entire data set (data frame) to f_i. The function called by f_i should be compatible with the type of data passed to it.

use_group

Optional grouping variable name to pass to imputation function if this supports group imputation.

group_level

A level of the framework to use for grouping indicators. This is only relevant if impute_by = "row". In that case, indicators will be split into their groups at the level specified by group_level, and imputation will be performed across rows of the group, rather than the whole data set. This can make more sense because indicators within a group are likely to be more similar.

normalise_first

Logical: if TRUE, each column is normalised using a min-max operation before imputation. By default this is FALSE unless impute_by = "row". See details.

out2

Either "coin" to return normalised data set back to the coin, or df to simply return a data frame.

write_to

Optional character string for naming the data set in the coin. Data will be written to .$Data[[write_to]]. Default is write_to == "Imputed".

...

arguments passed to or from other methods.

Value

An updated coin with imputed data set at .$Data[[write_to]]

Details

Clearly, the function f_i needs to be able to accept with the data class passed to it - if impute_by is "row" or "column" this will be a numeric vector, or if "df" it will be a data frame. Moreover, this function should return a vector or data frame identical to the vector/data frame passed to it except for NA values, which can be replaced. The function f_i is not required to replace all NA values.

When imputing row-wise, prior normalisation of the data is recommended. This is because imputation will use e.g. the mean of the unit values over all indicators (columns). If the indicators are on very different scales, the result will likely make no sense. If the indicators are normalised first, more sensible results can be obtained. There are two options to pre-normalise: first is by setting normalise_first = TRUE - this is anyway the default if impute_by = "row". In this case, you also need to supply a vector of directions. The data will then be normalised using a min-max approach before imputation, followed by the inverse operation to return the data to the original scales.

Another approach which gives more control is to simply run Normalise() first, and work with the normalised data from that point onwards. In that case it is better to set normalise_first = FALSE, since by default if impute_by = "row" it will be set to TRUE.

Checks are made on the format of the data returned by imputation functions, to ensure the type and that non-NA values have not been inadvertently altered. This latter check is allowed a degree of tolerance for numerical precision, controlled by the sfigs argument. This is because if the data frame is normalised, and/or depending on the imputation function, there may be a very small differences. By default sfigs = 9, meaning that the non-NA values pre and post-imputation are compared to 9 significant figures.

See also documentation for Impute.data.frame() and Impute.numeric() which are called by this function.

Examples

#' # build coin
coin <- build_example_coin(up_to = "new_coin")
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

# impute raw data set using population groups
# output to data frame directly
Impute(coin, dset = "Raw", f_i = "i_mean_grp",
               use_group = "Pop_group", out2 = "df")
#>    uCode      LPI   Flights      Ship Bord        Elec     Gas ConSpeed  Cov4G
#> 1    AUS 3.793385  36.05498 14.004198    0   0.0000000  1.1400 11.10000  94.00
#> 2    AUT 4.097985  29.01725  0.000000   35  35.3697298  0.2730 14.10000  98.00
#> 3    BEL 4.108538  31.88546 20.567121   48  26.5330467 36.1000 16.30000  99.89
#> 4    BGD 2.663902   4.27955  9.698165   16   1.3316167  0.1440 11.65000  65.00
#> 5    BGR 2.807685   9.23588  7.919366   18  11.2775842  0.3120 15.50000  56.73
#> 6    BRN 2.870492   2.01900  7.492593    2   0.0000000  6.0400 13.65455  80.00
#> 7    CHE 3.987158  51.78846  0.000000   41  68.7732872  3.0500 21.70000  98.00
#> 8    CHN 3.661104 114.20080 21.171976   32   6.4363058  9.4600  7.60000  85.00
#> 9    CYP 2.999061   8.75467 11.689495    0   0.4393643  0.0290  6.90000  60.00
#> 10   CZE 3.674309  15.30953  0.000000   35  46.7581787  2.7000 16.90000  98.70
#> 11   DEU 4.225967 174.35880 20.501262  105 110.1795628 48.7000 15.30000  95.70
#> 12   DNK 3.815794  32.77050 14.407080    3  20.1033650  0.4180 20.10000  99.99
#> 13   ESP 3.727412 170.96280 19.328835   26  24.1796233  3.2900 15.50000  91.30
#> 14   EST 3.363489   3.12946  9.171997    6   9.4990718  0.0667 11.60000 100.00
#> 15   FIN 3.920745  18.85806 10.044155   18  24.4142605  0.3800 20.50000  99.90
#> 16   FRA 3.900953  97.63489 19.536115   75  78.5259303 29.9000 10.80000  80.00
#> 17   GBR 4.069669 210.82440 20.800425   13  25.7837077 42.5000 16.90000  97.80
#> 18   GRC 3.239516  34.83849 15.679508   12   2.4308602  0.1390  7.90000  83.00
#> 19   HRV 3.160829   9.24529 12.440452   41  19.5283620  0.4220  8.60000  98.00
#> 20   HUN 3.428968  13.01196  0.000000   28  20.5432412  0.4190 14.80000  97.30
#>         Goods  Services    FDI     PRemit    ForPort Embs IGOs   UNVote
#> 1   288.48930 108.66450 20.900 14.7103856 2188.31100   82  196 38.46245
#> 2   278.42640 108.12730  5.000  5.0690118  808.68320   88  227 42.63920
#> 3   597.87230 216.31460  5.710 13.3545270 1574.30800   84  248 43.00308
#> 4    56.24166  10.07891  3.010  7.4693891    8.48896   52  145 38.60601
#> 5    42.82515  12.97674  1.350  1.0395414   15.50880   67  209 42.95986
#> 6     7.23391   2.11464  0.130  0.7604628 1642.16572   41   85 38.50109
#> 7   402.86490 207.36370 12.200  9.4893412 2615.08900   82  222 42.64127
#> 8  1713.61900 657.10900 75.600 26.5510247 1541.03500  100  193 37.13661
#> 9     8.76681  15.23712  1.230  0.4772705   40.76228   43  172 42.34705
#> 10  274.13650  43.46382  3.880  4.6915576  107.71030   84  201 42.22506
#> 11 1919.19400 578.42640 47.000 28.6815953 6386.91800  100  309 42.62615
#> 12  146.67710 113.67530  9.100  2.1877554 1021.27800   77  259 42.79140
#> 13  447.12290 197.19810 17.100 16.7522764 2040.57100   89  280 43.10103
#> 14   28.24110  10.22780  0.580  0.5888337   17.41899   46  194 42.87997
#> 15  101.57500  53.80751  6.030  1.5102145  747.62680   74  269 42.74629
#> 16  849.33030 471.29310 30.900 30.2099053 6745.32800  100  329 40.42295
#> 17  778.90520 518.22560 55.500 20.6965329 8206.97800  100  285 39.40118
#> 18   56.62702  38.17887  0.883  0.7630075  185.10480   83  216 42.69301
#> 19   28.36795  17.35676  0.387  1.5624119   16.88908   62  197 42.74966
#> 20  173.61590  39.11952  3.210  4.1655816   72.52570   79  212 42.53351
#>    CostImpEx Tariff      TBTs TIRcon RTAs Visa  StMob Research      Pat
#> 1      364.0   1.17  205.0000      0   14   68 268.00    46761  229.800
#> 2        0.0   1.60 1144.0000      1   30   80  69.80    14432  519.100
#> 3        0.0   1.60 1348.0000      1   30   80  42.30    20176  655.800
#> 4      595.0  10.53  526.9091      0   12   45  50.40     2084  627.160
#> 5       52.0   1.60 1140.0000      1   30   79  28.80     1854   12.200
#> 6      140.0   0.50    2.0000      0   15   90   3.50      229   41.925
#> 7      150.0   0.00  300.0000      1   33   80  49.80    30056 1315.700
#> 8      255.5   3.54 1230.0000      1   17   11 445.00    95919  927.800
#> 9      100.0   1.60 1141.0000      1   30   79  29.40     1337    7.400
#> 10       0.0   1.60 1456.0000      1   30   79  47.90     9190  112.600
#> 11      45.0   1.60 1163.0000      1   30   82 233.00    78353 2771.700
#> 12       0.0   1.60 1393.0000      1   30   81  33.20    15986  306.000
#> 13       0.0   1.60 1212.0000      1   30   81  51.00    39777  334.700
#> 14       0.0   1.60 1153.0000      1   30   79   5.99     1594   14.400
#> 15      70.0   1.60 1215.0000      1   30   81  22.00    10303  412.500
#> 16       0.0   1.60 1385.0000      1   30   81 152.00    56677 1469.000
#> 17      25.0   1.60 1187.0000      1   30   81 326.00    96337 1313.700
#> 18      30.0   1.60 1140.0000      1   30   80  48.60     8020   43.400
#> 19       0.0   1.60 1179.0000      1   30   78   5.05     2342   15.100
#> 20       0.0   1.60 1173.0000      1   30   79  26.20     5180  109.500
#>    CultServ CultGood Tourist MigStock       Lang       Renew  PrimEner
#> 1  2.000980  6.80700   8.263  5.66849 19.0054375  9.18050296 123.01003
#> 2  1.527580  8.93200  28.121  1.30365 14.1904160 34.39499175  84.92785
#> 3  2.387870 12.69800   7.481  1.32919 10.6908518  9.20164053 113.76871
#> 4  0.014720 22.26911   0.125  5.08139  0.1774405 34.74706901  74.74533
#> 5  0.119910  0.88000   8.252  0.69462  6.2899776 17.65007793 151.98183
#> 6  1.155337  0.36400   0.219  0.14041  6.2416879  0.01493504 113.44995
#> 7  1.159760 28.96500   9.205  2.12634 13.6355487 25.29198368  53.47854
#> 8  2.886240 74.46800  59.270  4.30982  0.7396357 12.41335257 175.31089
#> 9  0.137000  0.52900   3.187  0.23838 12.3211798  9.94206602  77.69450
#> 10 0.322300  8.80700   9.321  1.10637  7.7287677 14.82856195 134.44935
#> 11 4.836030 47.16100  35.555 10.92850 15.2613495 14.20625270  86.76327
#> 12 1.988160  4.91200  10.781  0.54405 19.2699694 33.17027717  63.75411
#> 13 1.592018 11.75900  75.315  3.34468  5.1584675 16.25409774  79.00655
#> 14 0.102750  0.74500   3.147  0.32117 12.2115280 27.47683347 169.38370
#> 15 0.588780  1.85800   2.789  0.46894 15.0435334 43.23526259 159.24526
#> 16 7.019050 41.02800  82.570  4.37339 11.5031818 13.49926232  97.46473
#> 17 9.569960 35.13900  35.814  8.91445 19.3521302  8.71168557  72.59383
#> 18 0.474830  1.62600  24.799  1.06705 10.5901399 17.17004755  88.19577
#> 19 0.283310  0.73700  13.809  0.54069 11.9764149 33.12746664  94.25010
#> 20 1.510710  2.61400   5.302  0.84562  5.2249358 15.55936205  95.79304
#>          CO2    MatCon    Forest    Poverty Palma TertGrad FreePress   TolMin
#> 1  15.370138 38.381073 6.3295294  0.3000000 1.400 29.78425        22 3.600000
#> 2   6.873713 15.763616 3.0382810  0.7000000 1.110 13.08167        22 4.800000
#> 3   8.328160 15.884942 5.7098277  0.0000000 0.960 31.70948        12 4.600000
#> 4   0.459142  2.577148 5.2605370 19.6000000 1.260 14.60314        62 8.400000
#> 5   5.871616 18.704688 1.5483901  1.5000000 1.490 24.47098        42 5.100000
#> 6  22.124701 23.331756 3.2269754  0.3363636 1.239 25.99135        76 5.600000
#> 7   4.311563 13.520983 0.9916749  0.0000000 1.200 36.94929        13 3.600000
#> 8   7.543908 23.647067 3.7251454  1.4000000 2.100  3.57765        87 7.600000
#> 9   5.260352 24.426326 1.7662068  0.0000000 1.400 25.54672        23 6.000000
#> 10  9.165978 17.115225 4.9847983  0.0000000 0.920 19.34180        21 5.100000
#> 11  8.889370 15.015616 2.3703489  0.0000000 1.100 24.74205        20 4.900000
#> 12  5.935712 16.651923 5.3043935  0.2000000 1.040 30.38373        12 4.400000
#> 13  5.033824 11.998735 6.9265991  1.0000000 1.460 29.29108        28 5.800000
#> 14 14.848819 32.980617 9.4062132  0.5000000 1.260 37.60498        16 7.536316
#> 15  8.660721 24.605785 8.5039377  0.0000000 0.950 22.50554        12 1.500000
#> 16  4.572088 11.827461 3.7879808  0.0000000 1.310 17.66307        26 7.000000
#> 17  6.497440  8.495755 5.4662193  0.2000000 1.250 29.33262        25 6.400000
#> 18  6.180337 11.624611 1.8997660  1.5000000 1.550 22.92907        44 5.100000
#> 19  3.973805  9.791906 1.7280976  0.7000000 1.220 18.25773        41 5.500000
#> 20  4.265575 16.923169 5.8471809  0.5000000 1.130 20.67309        44 4.500000
#>    NGOs CPI    FemLab WomParl PubDebt  PrivDebt     GDPGrow    RDExp     NEET
#> 1   311  77 0.8592661 28.6667   37.64 202.50016  0.34888036 2.202070  8.70000
#> 2   301  75 0.8872138 30.6011   86.19 141.28513  2.18677558 3.071810  6.50000
#> 3  1824  75 0.8699724 38.0000  106.05 225.72427  1.36616464 2.456680  9.30000
#> 4    16  28 0.5414029 20.2857   31.67  40.81502  6.16478747 1.196272 27.40000
#> 5    28  43 0.8927631 19.1667   26.30 116.98938  4.32022261 0.956570 15.30000
#> 6     1  62 0.8618006  9.0909    2.81 189.43663  0.02851029 1.124665 17.20000
#> 7   741  85 0.8936418 32.5000   45.66 237.77122 -0.01982295 2.966350  6.50000
#> 8   117  41 0.8359555 23.7030   42.92 193.43419  6.30396712 2.065580 16.09091
#> 9    19  57 0.8633334 17.8571  108.87 345.59967  0.00000000 0.455850 16.10000
#> 10   55  57 0.8119483 20.0000   40.31  73.21357  4.04300124 1.948650  6.30000
#> 11  986  81 0.8844910 36.9841   70.99 105.91991  1.79471020 2.877490  6.30000
#> 12  208  88 0.9270000 37.4302   45.53 223.23741  1.50309097 3.013870  7.00000
#> 13  341  57 0.8631487 39.1429   99.26 165.79129  2.85707386 1.219610 13.30000
#> 14   21  71 0.9000751 26.7327    9.74 126.58231  4.87884628 1.495020  9.40000
#> 15  118  85 0.9651910 42.0000   62.51 181.23815  2.33629470 2.904740  9.40000
#> 16 1071  70 0.8953486 25.8232   96.14 188.68200  1.42667764 2.231350 11.50000
#> 17 1582  82 0.8697419 30.0000   88.98 169.14574  1.12929130 1.703040 10.30000
#> 18  102  48 0.7784910 18.3333  176.94 124.65858  1.49779228 0.956730 15.30000
#> 19   30  49 0.8664952 19.8675   86.67 130.34435  3.99021292 0.854390 15.40000
#> 20   62  45 0.8249007 10.0503   75.33  92.51395  4.33844641 1.377620 11.00000
#>  [ reached 'max' / getOption("max.print") -- omitted 31 rows ]