Impute a data frame

Impute a data frame using any function, either column-wise, row-wise or by the whole data frame in one shot.

# S3 method for data.frame
Impute(
  x,
  f_i = NULL,
  f_i_para = NULL,
  impute_by = "column",
  normalise_first = NULL,
  directions = NULL,
  ...
)

Arguments

x

A data frame with only numeric columns.

f_i

A function to use for imputation. By default, imputation is performed by simply substituting the mean of non-NA values for each column at a time.

f_i_para

Any additional parameters to pass to f_i, apart from x

impute_by

Specifies how to impute: if "column", passes each column separately as a numerical vector to f_i; if "row", passes each row separately; and if "df" passes the entire data frame to f_i. The function called by f_i should be compatible with the type of data passed to it.

normalise_first

Logical: if TRUE, each column is normalised using a min-max operation before imputation. By default this is FALSE unless impute_by = "row". See details.

directions

A vector of directions: either -1 or 1 to indicate the direction of each column of x - this is only used if normalise_first = TRUE. See details.

...

arguments passed to or from other methods.

Value

An imputed data frame

Details

This function only accepts data frames with all numeric columns. It imputes any NAs in the data frame by invoking the function f_i and any optional arguments f_i_para on each column at a time (if impute_by = "column"), or on each row at a time (if impute_by = "row"), or by passing the entire data frame to f_i if impute_by = "df".

Clearly, the function f_i needs to be able to accept with the data class passed to it - if impute_by is "row" or "column" this will be a numeric vector, or if "df" it will be a data frame. Moreover, this function should return a vector or data frame identical to the vector/data frame passed to it except for NA values, which can be replaced. The function f_i is not required to replace all NA values.

When imputing row-wise, prior normalisation of the data is recommended. This is because imputation will use e.g. the mean of the unit values over all indicators (columns). If the indicators are on very different scales, the result will likely make no sense. If the indicators are normalised first, more sensible results can be obtained. There are two options to pre-normalise: first is by setting normalise_first = TRUE - this is anyway the default if impute_by = "row". In this case, you also need to supply a vector of directions. The data will then be normalised using a min-max approach before imputation, followed by the inverse operation to return the data to the original scales.

Another approach which gives more control is to simply run Normalise() first, and work with the normalised data from that point onwards. In that case it is better to set normalise_first = FALSE, since by default if impute_by = "row" it will be set to TRUE.

Checks are made on the format of the data returned by imputation functions, to ensure the type and that non-NA values have not been inadvertently altered. This latter check is allowed a degree of tolerance for numerical precision, controlled by the sfigs argument. This is because if the data frame is normalised, and/or depending on the imputation function, there may be a very small differences. By default sfigs = 9, meaning that the non-NA values pre and post-imputation are compared to 9 significant figures.

Examples

# a df of random numbers
X <- as.data.frame(matrix(runif(50), 10, 5))

# introduce NAs (2 in 3 of 5 cols)
X[sample(1:10, 2), 1] <- NA
X[sample(1:10, 2), 3] <- NA
X[sample(1:10, 2), 5] <- NA

# impute using column mean
Impute(X, f_i = "i_mean")
#>           V1         V2        V3        V4        V5
#> 1  0.1497639 0.52562715 0.4087492 0.7811421 0.6241937
#> 2  0.1511259 0.97964470 0.3694333 0.6596783 0.9998536
#> 3  0.4819219 0.98923273 0.7469984 0.2349049 0.7604280
#> 4  0.6677410 0.52659494 0.4734876 0.2457475 0.1203534
#> 5  0.6695454 0.09680426 0.1229599 0.6946503 0.6295385
#> 6  0.4819219 0.56856717 0.7196786 0.2436740 0.6295385
#> 7  0.1899729 0.14015188 0.8602717 0.5614376 0.9552747
#> 8  0.4772996 0.59003401 0.2409119 0.7965775 0.3987586
#> 9  0.7766469 0.93166719 0.4734876 0.6214333 0.1943477
#> 10 0.7732800 0.19964817 0.3188979 0.4702918 0.9830982

# impute using row median (no normalisation)
Impute(X, f_i = "i_median", impute_by = "row",
       normalise_first = FALSE)
#>           V1         V2        V3        V4        V5
#> 1  0.1497639 0.52562715 0.4087492 0.7811421 0.6241937
#> 2  0.1511259 0.97964470 0.3694333 0.6596783 0.9998536
#> 3  0.7537132 0.98923273 0.7469984 0.2349049 0.7604280
#> 4  0.6677410 0.52659494 0.3861712 0.2457475 0.1203534
#> 5  0.6695454 0.09680426 0.1229599 0.6946503 0.3962526
#> 6  0.5685672 0.56856717 0.7196786 0.2436740 0.5685672
#> 7  0.1899729 0.14015188 0.8602717 0.5614376 0.9552747
#> 8  0.4772996 0.59003401 0.2409119 0.7965775 0.3987586
#> 9  0.7766469 0.93166719 0.6990401 0.6214333 0.1943477
#> 10 0.7732800 0.19964817 0.3188979 0.4702918 0.9830982