Get subsets of indicator data

A helper function to retrieve a named data set from coin or purse objects. See individual method documentation:

Usage

get_data(x, ...)

Arguments

x

A coin or purse

...

Arguments passed to methods

Value

Data frame of indicator data, indexed also by time if input is a purse.

Details

get_data.coin()

get_data.purse()

This function replaces the now-defunct getIn() from COINr < v1.0.

Examples

# see individual method documentation