Gets a named data set and performs checks

A helper function to retrieve a named data set from the coin object. Also performs input checks at the same time.

Usage

# S3 method for coin
get_dset(x, dset, also_get = NULL, ...)

Arguments

x

A coin class object

dset

A character string corresponding to a named data set within .$Data. E.g. "Raw".

also_get

A character vector specifying any columns to attach to the data set that are not indicators or aggregates. These will be e.g. uName, groups, denominators or columns labelled as "Other" in iMeta. These columns are stored in .$Meta$Unit to avoid repetition. Set also_get = "all" to attach all columns, or set also_get = "none" to return only numeric columns, i.e. no uCode column.

...

arguments passed to or from other methods.

Value

Data frame of indicator data.

Details

If also_get is not specified, this will return the indicator columns with the uCode identifiers in the first column. Optionally, also_get can be specified to attach other metadata columns, or to only return the numeric (indicator) columns with no identifiers. This latter option might be useful for e.g. examining correlations.

Examples

# build example coin, just up to raw dset for speed
coin <- build_example_coin(up_to = "new_coin", quietly = TRUE)

# retrieve raw data set with added cols
get_dset(coin, dset = "Raw", also_get = c("uName", "GDP_group"))
#>    uCode             uName GDP_group      LPI   Flights      Ship Bord
#> 1    AUS         Australia        XL 3.793385  36.05498 14.004198    0
#> 2    AUT           Austria         L 4.097985  29.01725  0.000000   35
#> 3    BEL           Belgium         L 4.108538  31.88546 20.567121   48
#> 4    BGD        Bangladesh         M 2.663902   4.27955  9.698165   16
#> 5    BGR          Bulgaria         S 2.807685   9.23588  7.919366   18
#> 6    BRN Brunei Darussalam         S 2.870492   2.01900  7.492593    2
#> 7    CHE       Switzerland         L 3.987158  51.78846  0.000000   41
#> 8    CHN             China        XL 3.661104 114.20080 21.171976   32
#> 9    CYP            Cyprus         S 2.999061   8.75467 11.689495    0
#> 10   CZE    Czech Republic         M 3.674309  15.30953  0.000000   35
#> 11   DEU           Germany        XL 4.225967 174.35880 20.501262  105
#> 12   DNK           Denmark         L 3.815794  32.77050 14.407080    3
#> 13   ESP             Spain        XL 3.727412 170.96280 19.328835   26
#> 14   EST           Estonia         S 3.363489   3.12946  9.171997    6
#> 15   FIN           Finland         M 3.920745  18.85806 10.044155   18
#> 16   FRA            France        XL 3.900953  97.63489 19.536115   75
#> 17   GBR    United Kingdom        XL 4.069669 210.82440 20.800425   13
#> 18   GRC            Greece         M 3.239516  34.83849 15.679508   12
#> 19   HRV           Croatia         S 3.160829   9.24529 12.440452   41
#>           Elec     Gas ConSpeed  Cov4G      Goods  Services    FDI     PRemit
#> 1    0.0000000  1.1400     11.1  94.00  288.48930 108.66450 20.900 14.7103856
#> 2   35.3697298  0.2730     14.1  98.00  278.42640 108.12730  5.000  5.0690118
#> 3   26.5330467 36.1000     16.3  99.89  597.87230 216.31460  5.710 13.3545270
#> 4    1.3316167  0.1440       NA  65.00   56.24166  10.07891  3.010  7.4693891
#> 5   11.2775842  0.3120     15.5  56.73   42.82515  12.97674  1.350  1.0395414
#> 6    0.0000000  6.0400       NA  80.00    7.23391   2.11464  0.130  0.7604628
#> 7   68.7732872  3.0500     21.7  98.00  402.86490 207.36370 12.200  9.4893412
#> 8    6.4363058  9.4600      7.6  85.00 1713.61900 657.10900 75.600 26.5510247
#> 9    0.4393643  0.0290      6.9  60.00    8.76681  15.23712  1.230  0.4772705
#> 10  46.7581787  2.7000     16.9  98.70  274.13650  43.46382  3.880  4.6915576
#> 11 110.1795628 48.7000     15.3  95.70 1919.19400 578.42640 47.000 28.6815953
#> 12  20.1033650  0.4180     20.1  99.99  146.67710 113.67530  9.100  2.1877554
#> 13  24.1796233  3.2900     15.5  91.30  447.12290 197.19810 17.100 16.7522764
#> 14   9.4990718  0.0667     11.6 100.00   28.24110  10.22780  0.580  0.5888337
#> 15  24.4142605  0.3800     20.5  99.90  101.57500  53.80751  6.030  1.5102145
#> 16  78.5259303 29.9000     10.8  80.00  849.33030 471.29310 30.900 30.2099053
#> 17  25.7837077 42.5000     16.9  97.80  778.90520 518.22560 55.500 20.6965329
#> 18   2.4308602  0.1390      7.9  83.00   56.62702  38.17887  0.883  0.7630075
#> 19  19.5283620  0.4220      8.6  98.00   28.36795  17.35676  0.387  1.5624119
#>       ForPort Embs IGOs   UNVote CostImpEx Tariff TBTs TIRcon RTAs Visa  StMob
#> 1  2188.31100   82  196 38.46245     364.0   1.17  205      0   14   68 268.00
#> 2   808.68320   88  227 42.63920       0.0   1.60 1144      1   30   80  69.80
#> 3  1574.30800   84  248 43.00308       0.0   1.60 1348      1   30   80  42.30
#> 4     8.48896   52  145 38.60601     595.0  10.53   NA      0   12   45  50.40
#> 5    15.50880   67  209 42.95986      52.0   1.60 1140      1   30   79  28.80
#> 6          NA   41   85 38.50109     140.0   0.50    2      0   15   90   3.50
#> 7  2615.08900   82  222 42.64127     150.0   0.00  300      1   33   80  49.80
#> 8  1541.03500  100  193 37.13661     255.5   3.54 1230      1   17   11 445.00
#> 9    40.76228   43  172 42.34705     100.0   1.60 1141      1   30   79  29.40
#> 10  107.71030   84  201 42.22506       0.0   1.60 1456      1   30   79  47.90
#> 11 6386.91800  100  309 42.62615      45.0   1.60 1163      1   30   82 233.00
#> 12 1021.27800   77  259 42.79140       0.0   1.60 1393      1   30   81  33.20
#> 13 2040.57100   89  280 43.10103       0.0   1.60 1212      1   30   81  51.00
#> 14   17.41899   46  194 42.87997       0.0   1.60 1153      1   30   79   5.99
#> 15  747.62680   74  269 42.74629      70.0   1.60 1215      1   30   81  22.00
#> 16 6745.32800  100  329 40.42295       0.0   1.60 1385      1   30   81 152.00
#> 17 8206.97800  100  285 39.40118      25.0   1.60 1187      1   30   81 326.00
#> 18  185.10480   83  216 42.69301      30.0   1.60 1140      1   30   80  48.60
#> 19   16.88908   62  197 42.74966       0.0   1.60 1179      1   30   78   5.05
#>    Research    Pat CultServ CultGood Tourist MigStock       Lang       Renew
#> 1     46761  229.8  2.00098    6.807   8.263  5.66849 19.0054375  9.18050296
#> 2     14432  519.1  1.52758    8.932  28.121  1.30365 14.1904160 34.39499175
#> 3     20176  655.8  2.38787   12.698   7.481  1.32919 10.6908518  9.20164053
#> 4      2084     NA  0.01472       NA   0.125  5.08139  0.1774405 34.74706901
#> 5      1854   12.2  0.11991    0.880   8.252  0.69462  6.2899776 17.65007793
#> 6       229     NA       NA    0.364   0.219  0.14041  6.2416879  0.01493504
#> 7     30056 1315.7  1.15976   28.965   9.205  2.12634 13.6355487 25.29198368
#> 8     95919  927.8  2.88624   74.468  59.270  4.30982  0.7396357 12.41335257
#> 9      1337    7.4  0.13700    0.529   3.187  0.23838 12.3211798  9.94206602
#> 10     9190  112.6  0.32230    8.807   9.321  1.10637  7.7287677 14.82856195
#> 11    78353 2771.7  4.83603   47.161  35.555 10.92850 15.2613495 14.20625270
#> 12    15986  306.0  1.98816    4.912  10.781  0.54405 19.2699694 33.17027717
#> 13    39777  334.7       NA   11.759  75.315  3.34468  5.1584675 16.25409774
#> 14     1594   14.4  0.10275    0.745   3.147  0.32117 12.2115280 27.47683347
#> 15    10303  412.5  0.58878    1.858   2.789  0.46894 15.0435334 43.23526259
#> 16    56677 1469.0  7.01905   41.028  82.570  4.37339 11.5031818 13.49926232
#> 17    96337 1313.7  9.56996   35.139  35.814  8.91445 19.3521302  8.71168557
#> 18     8020   43.4  0.47483    1.626  24.799  1.06705 10.5901399 17.17004755
#> 19     2342   15.1  0.28331    0.737  13.809  0.54069 11.9764149 33.12746664
#>     PrimEner       CO2    MatCon    Forest Poverty Palma TertGrad FreePress
#> 1  123.01003 15.370138 38.381073 6.3295294     0.3  1.40 29.78425        22
#> 2   84.92785  6.873713 15.763616 3.0382810     0.7  1.11 13.08167        22
#> 3  113.76871  8.328160 15.884942 5.7098277     0.0  0.96 31.70948        12
#> 4   74.74533  0.459142  2.577148 5.2605370    19.6  1.26       NA        62
#> 5  151.98183  5.871616 18.704688 1.5483901     1.5  1.49 24.47098        42
#> 6  113.44995 22.124701 23.331756 3.2269754      NA    NA       NA        76
#> 7   53.47854  4.311563 13.520983 0.9916749     0.0  1.20 36.94929        13
#> 8  175.31089  7.543908 23.647067 3.7251454     1.4  2.10  3.57765        87
#> 9   77.69450  5.260352 24.426326 1.7662068     0.0  1.40 25.54672        23
#> 10 134.44935  9.165978 17.115225 4.9847983     0.0  0.92 19.34180        21
#> 11  86.76327  8.889370 15.015616 2.3703489     0.0  1.10 24.74205        20
#> 12  63.75411  5.935712 16.651923 5.3043935     0.2  1.04 30.38373        12
#> 13  79.00655  5.033824 11.998735 6.9265991     1.0  1.46 29.29108        28
#> 14 169.38370 14.848819 32.980617 9.4062132     0.5  1.26 37.60498        16
#> 15 159.24526  8.660721 24.605785 8.5039377     0.0  0.95 22.50554        12
#> 16  97.46473  4.572088 11.827461 3.7879808     0.0  1.31 17.66307        26
#> 17  72.59383  6.497440  8.495755 5.4662193     0.2  1.25 29.33262        25
#> 18  88.19577  6.180337 11.624611 1.8997660     1.5  1.55 22.92907        44
#> 19  94.25010  3.973805  9.791906 1.7280976     0.7  1.22 18.25773        41
#>      TolMin NGOs CPI    FemLab WomParl PubDebt  PrivDebt     GDPGrow   RDExp
#> 1  3.600000  311  77 0.8592661 28.6667   37.64 202.50016  0.34888036 2.20207
#> 2  4.800000  301  75 0.8872138 30.6011   86.19 141.28513  2.18677558 3.07181
#> 3  4.600000 1824  75 0.8699724 38.0000  106.05 225.72427  1.36616464 2.45668
#> 4  8.400000   16  28 0.5414029 20.2857   31.67  40.81502  6.16478747      NA
#> 5  5.100000   28  43 0.8927631 19.1667   26.30 116.98938  4.32022261 0.95657
#> 6  5.600000    1  62        NA  9.0909    2.81        NA  0.02851029      NA
#> 7  3.600000  741  85 0.8936418 32.5000   45.66 237.77122 -0.01982295 2.96635
#> 8  7.600000  117  41 0.8359555 23.7030   42.92 193.43419  6.30396712 2.06558
#> 9  6.000000   19  57 0.8633334 17.8571  108.87 345.59967  0.00000000 0.45585
#> 10 5.100000   55  57 0.8119483 20.0000   40.31  73.21357  4.04300124 1.94865
#> 11 4.900000  986  81 0.8844910 36.9841   70.99 105.91991  1.79471020 2.87749
#> 12 4.400000  208  88 0.9270000 37.4302   45.53 223.23741  1.50309097 3.01387
#> 13 5.800000  341  57 0.8631487 39.1429   99.26 165.79129  2.85707386 1.21961
#> 14 7.536316   21  71 0.9000751 26.7327    9.74 126.58231  4.87884628 1.49502
#> 15 1.500000  118  85 0.9651910 42.0000   62.51 181.23815  2.33629470 2.90474
#> 16 7.000000 1071  70 0.8953486 25.8232   96.14 188.68200  1.42667764 2.23135
#> 17 6.400000 1582  82 0.8697419 30.0000   88.98 169.14574  1.12929130 1.70304
#> 18 5.100000  102  48 0.7784910 18.3333  176.94 124.65858  1.49779228 0.95673
#> 19 5.500000   30  49 0.8664952 19.8675   86.67 130.34435  3.99021292 0.85439
#>    NEET
#> 1   8.7
#> 2   6.5
#> 3   9.3
#> 4  27.4
#> 5  15.3
#> 6  17.2
#> 7   6.5
#> 8    NA
#> 9  16.1
#> 10  6.3
#> 11  6.3
#> 12  7.0
#> 13 13.3
#> 14  9.4
#> 15  9.4
#> 16 11.5
#> 17 10.3
#> 18 15.3
#> 19 15.4
#>  [ reached 'max' / getOption("max.print") -- omitted 32 rows ]