Normalise as distance to target

A measure of the distance of each value of x to a specified target. The formula is:

n_dist2targ(x, targ, cap_max = FALSE)

Arguments

x

A numeric vector

targ

An target value

cap_max

If TRUE, any value of x that exceeds targ will be assigned a score of 1, otherwise will have a score greater than 1.

Value

Numeric vector

Details

$$ 1 - (x_{targ} - x)/(x_{targ} - x_{min}) $$

Values exceeding x_targ can be optionally capped at 1 if cap_max = TRUE.

Examples

x <- runif(20)
n_dist2targ(x, 0.8, cap_max = TRUE)
#>  [1] 0.10778656 0.67844529 0.93064812 0.70720715 0.44798245 0.35225254
#>  [7] 0.64772987 1.00000000 0.16908825 0.97147303 0.64132345 0.16067846
#> [13] 0.00000000 0.39332089 0.01449623 0.16934249 1.00000000 0.75895383
#> [19] 0.05241941 0.71003254