Plot sensitivity indices

Plots sensitivity indices as bar or pie charts.

Usage

plot_sensitivity(SAresults, ptype = "bar")

Arguments

SAresults

A list of sensitivity/uncertainty analysis results from plot_sensitivity().

ptype

Type of plot to generate - either "bar", "pie" or "box".

Value

A plot of sensitivity indices generated by ggplot2.

Details

To use this function you first need to run get_sensitivity(). Then enter the resulting list as the SAresults argument here.

See vignette("sensitivity").

This function replaces the now-defunct plotSA() from COINr < v1.0.

See also

Examples

# for examples, see `vignette("sensitivity")`
# (this is because package examples are run automatically and sensitivity analysis
# can take a few minutes to run at realistic settings)