Normalise a numeric vector

Normalise a numeric vector using a specified function f_n, with possible reversal of direction using direction.

Usage

# S3 method for numeric
Normalise(x, f_n = NULL, f_n_para = NULL, direction = 1, ...)

Arguments

x

Object to be normalised

f_n

The normalisation method, specified as string which refers to a function of the form f_n(x, npara). See details. Defaults to "n_minmax" which is the min-max function.

f_n_para

Supporting list of arguments for f_n. This is required to be a list.

direction

If direction = -1 the highest values of x will correspond to the lowest values of the normalised x. Else if direction = 1 the direction of x in unaltered.

...

arguments passed to or from other methods.

Value

A normalised numeric vector

Details

Normalisation is specified using the f_n and f_n_para arguments. In these, f_n should be a character string which is the name of a normalisation function. For example, f_n = "n_minmax" calls the n_minmax() function. f_n_para is a list of any further arguments to f_n. This means that any function can be passed to Normalise(), as long as its first argument is x, a numeric vector, and it returns a numeric vector of the same length. See n_minmax() for an example.

f_n_para is required to be a named list. So e.g. if we define a function f1(x, arg1, arg2) then we should specify f_n = "f1", and f_n_para = list(arg1 = val1, arg2 = val2), where val1 and val2 are the values assigned to the arguments arg1 and arg2 respectively.

See also vignette("normalise") for more details.

Examples

# example vector
x <- runif(10)

# normalise using distance to reference (5th data point)
x_norm <- Normalise(x, f_n = "n_dist2ref", f_n_para = list(iref = 5))

# view side by side
data.frame(x, x_norm)
#>             x     x_norm
#> 1  0.20027021  1.9763620
#> 2  0.41662124  4.4084964
#> 3  0.21947003  2.1921990
#> 4  0.02446251  0.0000000
#> 5  0.11341772  1.0000000
#> 6  0.26940341  2.7535306
#> 7  0.50750940  5.4302259
#> 8  0.36808953  3.8629218
#> 9  0.08576060  0.6890894
#> 10 0.98376003 10.7840509