Treat a data set in a coin for outliers

Operates a two-stage data treatment process, based on two data treatment functions, and a pass/fail function which detects outliers. This function is set up to allow any functions to be passed as the data treatment functions (f1 and f2), as well as any function to be passed as the outlier detection function f_pass. The data set in the coin to be treated is specified by dset.

Usage

# S3 method for coin
Treat(
  x,
  dset,
  global_specs = NULL,
  indiv_specs = NULL,
  combine_treat = FALSE,
  out2 = "coin",
  write_to = NULL,
  write2log = TRUE,
  ...
)

Arguments

x

A coin

dset

A named data set available in .$Data

global_specs

Default specifications

indiv_specs

Individual specifications

combine_treat

By default, if f1 fails to pass f_pass, then f2 is applied to the original x, rather than the treated output of f1. If combine_treat = TRUE, f2 will instead be applied to the output of f1, so the two treatments will be combined.

out2

The type of function output: either "coin" to return an updated coin, or "list" to return a list with treated data and treatment details.

write_to

If specified, writes the aggregated data to .$Data[[write_to]]. Default write_to = "Treated".

write2log

Logical: if FALSE, the arguments of this function are not written to the coin log, so this function will not be invoked when regenerating. Recommend to keep TRUE unless you have a good reason to do otherwise.

...

arguments passed to or from other methods.

Value

An updated coin with a new data set .Data$Treated added, plus analysis information in .$Analysis$Treated.

Details

The arrangement of this function is inspired by a fairly standard data treatment process applied to indicators, which consists of checking skew and kurtosis, then if the criteria are not met, applying Winsorisation up to a specified limit. Then if Winsorisation still does not bring skew and kurtosis within limits, applying a nonlinear transformation such as log or Box-Cox.

This function generalises this process by using the following general steps:

Check if variable passes or fails using f_pass

If f_pass returns FALSE, apply f1, else return x unmodified

Check again using *f_pass

If f_pass still returns FALSE, apply f2

Return the modified x as well as other information.

For the "typical" case described above f1 is a Winsorisation function, f2 is a nonlinear transformation and f_pass is a skew and kurtosis check. Parameters can be passed to each of these three functions in a named list, for example to specify a maximum number of points to Winsorise, or Box-Cox parameters, or anything else. The constraints are that:

All of f1, f2 and f_pass must follow the format function(x, f_para), where x is a numerical vector, and f_para is a list of other function parameters to be passed to the function, which is specified by f1_para for f1 and similarly for the other functions. If the function has no parameters other than x, then f_para can be omitted.

f1 and f2 should return either a list with .$x as the modified numerical vector, and any other information to be attached to the list, OR, simply x as the only output.

f_pass must return a logical value, where TRUE indicates that the x passes the criteria (and therefore doesn't need any (more) treatment), and FALSE means that it fails to meet the criteria.

See also vignette("treat").

Examples

# build example coin
coin <- build_example_coin(up_to = "new_coin")
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

# treat raw data set
coin <- Treat(coin, dset = "Raw")
#> Written data set to .$Data$Treated

# summary of treatment for each indicator
head(coin$Analysis$Treated$Dets_Table)
#>     iCode check_SkewKurt0.Pass check_SkewKurt0.Skew check_SkewKurt0.Kurt
#> 1     LPI                 TRUE           -0.3042681           -0.6567514
#> 2 Flights                FALSE            2.1032872            4.5088794
#> 3    Ship                 TRUE           -0.5756680           -0.6814795
#> 4    Bord                FALSE            2.1482360            5.7914905
#> 5    Elec                FALSE            2.2252736            5.7910268
#> 6     Gas                FALSE            2.8294486           10.3346494
#>   winsorise.nwin check_SkewKurt1.Pass check_SkewKurt1.Skew check_SkewKurt1.Kurt
#> 1             NA                   NA                   NA                   NA
#> 2              1                 TRUE             1.900658             3.336065
#> 3             NA                   NA                   NA                   NA
#> 4              1                 TRUE             1.899211             4.346298
#> 5              1                 TRUE             1.717744             2.586062
#> 6              1                 TRUE             1.602518             1.525576