Gets a named data set and performs checks

A helper function to retrieve a named data set from the coin object. Also performs input checks at the same time.

Usage

# S3 method for coin
get_dset(x, dset, also_get = NULL, ...)

Arguments

x

A coin class object

dset

A character string corresponding to a named data set within .$Data. E.g. "Raw".

also_get

A character vector specifying any columns to attach to the data set that are not indicators or aggregates. These will be e.g. uName, groups, denominators or columns labelled as "Other" in iMeta. These columns are stored in .$Meta$Unit to avoid repetition. Set also_get = "all" to attach all columns, or set also_get = "none" to return only numeric columns, i.e. no uCode column.

...

arguments passed to or from other methods.

Value

Data frame of indicator data.

Details

If also_get is not specified, this will return the indicator columns with the uCode identifiers in the first column. Optionally, also_get can be specified to attach other metadata columns, or to only return the numeric (indicator) columns with no identifiers. This latter option might be useful for e.g. examining correlations.

Examples

# build example coin, just up to raw dset for speed
coin <- build_example_coin(up_to = "new_coin", quietly = TRUE)

# retrieve raw data set with added cols
get_dset(coin, dset = "Raw", also_get = c("uName", "GDP_group"))
#>    uCode              uName GDP_group      LPI   Flights      Ship Bord
#> 1    AUS          Australia        XL 3.793385  36.05498 14.004198    0
#> 2    AUT            Austria         L 4.097985  29.01725  0.000000   35
#> 3    BEL            Belgium         L 4.108538  31.88546 20.567121   48
#> 4    BGD         Bangladesh         M 2.663902   4.27955  9.698165   16
#> 5    BGR           Bulgaria         S 2.807685   9.23588  7.919366   18
#> 6    BRN  Brunei Darussalam         S 2.870492   2.01900  7.492593    2
#> 7    CHE        Switzerland         L 3.987158  51.78846  0.000000   41
#> 8    CHN              China        XL 3.661104 114.20080 21.171976   32
#> 9    CYP             Cyprus         S 2.999061   8.75467 11.689495    0
#> 10   CZE     Czech Republic         M 3.674309  15.30953  0.000000   35
#> 11   DEU            Germany        XL 4.225967 174.35880 20.501262  105
#> 12   DNK            Denmark         L 3.815794  32.77050 14.407080    3
#> 13   ESP              Spain        XL 3.727412 170.96280 19.328835   26
#> 14   EST            Estonia         S 3.363489   3.12946  9.171997    6
#> 15   FIN            Finland         M 3.920745  18.85806 10.044155   18
#> 16   FRA             France        XL 3.900953  97.63489 19.536115   75
#> 17   GBR     United Kingdom        XL 4.069669 210.82440 20.800425   13
#> 18   GRC             Greece         M 3.239516  34.83849 15.679508   12
#> 19   HRV            Croatia         S 3.160829   9.24529 12.440452   41
#> 20   HUN            Hungary         M 3.428968  13.01196  0.000000   28
#> 21   IDN          Indonesia        XL 2.984537  34.68020 12.209998    2
#> 22   IND              India        XL 3.420043  25.26430 16.522118   35
#> 23   IRL            Ireland         L 3.794886  34.17721 10.575193   12
#> 24   ITA              Italy        XL 3.755414 110.27550 18.383778   24
#> 25   JPN              Japan        XL 3.970464  69.00941 17.261338    0
#> 26   KAZ         Kazakhstan         M 2.751998   4.02303  0.000000   46
#> 27   KHM           Cambodia         S 2.800590   9.52120  9.017474   13
#> 28   KOR              Korea        XL 3.717126  69.84827 20.427418    1
#> 29   LAO            Lao PDR         S 2.067254   3.07440  0.000000   20
#> 30   LTU          Lithuania         S 3.631688   5.37919  9.234349   22
#> 31   LUX         Luxembourg         S 4.219409   4.84458  0.000000   12
#> 32   LVA             Latvia         S 3.327107   6.77976  7.850937   17
#> 33   MLT              Malta         S 3.069256   6.75251 17.060552    0
#> 34   MMR            Myanmar         M 2.458571   6.69500  9.182403    8
#> 35   MNG           Mongolia         S 2.506056   0.98951  0.000000    9
#> 36   MYS           Malaysia         L 3.426307  53.33988 20.551337    9
#> 37   NLD        Netherlands         L 4.187530  63.59241 20.541548   32
#> 38   NOR             Norway         L 3.732163  25.64994  9.287889   19
#> 39   NZL        New Zealand         M 3.388000  13.37242 11.910746    0
#> 40   PAK           Pakistan         M 2.923219   2.21146 15.319999   43
#> 41   PHL        Philippines         L 2.856259  19.43838 11.621228    0
#> 42   POL             Poland         L 3.425877  33.79735 14.815517   62
#> 43   PRT           Portugal         M 3.409367  40.46484 16.198871   11
#> 44   ROU            Romania         M 2.993120  18.79894 12.689838   24
#> 45   RUS Russian Federation        XL 2.570864  34.06447 15.811485  122
#> 46   SGP          Singapore         L 4.143632  67.86156 20.276872    2
#> 47   SVK           Slovakia         M 3.336895   2.07397  0.000000   25
#> 48   SVN           Slovenia         S 3.184508   1.51736 13.380177   20
#> 49   SWE             Sweden         L 4.204593  31.74444 15.578093   18
#> 50   THA           Thailand         L 3.255100  73.92221 14.506399   17
#> 51   VNM            Vietnam         M 2.976629  28.52418 16.548761   25
#>            Elec      Gas ConSpeed  Cov4G      Goods  Services    FDI     PRemit
#> 1    0.00000000  1.14000     11.1  94.00  288.48930 108.66450 20.900 14.7103856
#> 2   35.36972983  0.27300     14.1  98.00  278.42640 108.12730  5.000  5.0690118
#> 3   26.53304667 36.10000     16.3  99.89  597.87230 216.31460  5.710 13.3545270
#> 4    1.33161667  0.14400       NA  65.00   56.24166  10.07891  3.010  7.4693891
#> 5   11.27758417  0.31200     15.5  56.73   42.82515  12.97674  1.350  1.0395414
#> 6    0.00000000  6.04000       NA  80.00    7.23391   2.11464  0.130  0.7604628
#> 7   68.77328717  3.05000     21.7  98.00  402.86490 207.36370 12.200  9.4893412
#> 8    6.43630583  9.46000      7.6  85.00 1713.61900 657.10900 75.600 26.5510247
#> 9    0.43936433  0.02900      6.9  60.00    8.76681  15.23712  1.230  0.4772705
#> 10  46.75817867  2.70000     16.9  98.70  274.13650  43.46382  3.880  4.6915576
#> 11 110.17956283 48.70000     15.3  95.70 1919.19400 578.42640 47.000 28.6815953
#> 12  20.10336500  0.41800     20.1  99.99  146.67710 113.67530  9.100  2.1877554
#> 13  24.17962333  3.29000     15.5  91.30  447.12290 197.19810 17.100 16.7522764
#> 14   9.49907183  0.06670     11.6 100.00   28.24110  10.22780  0.580  0.5888337
#> 15  24.41426050  0.38000     20.5  99.90  101.57500  53.80751  6.030  1.5102145
#> 16  78.52593033 29.90000     10.8  80.00  849.33030 471.29310 30.900 30.2099053
#> 17  25.78370767 42.50000     16.9  97.80  778.90520 518.22560 55.500 20.6965329
#> 18   2.43086017  0.13900      7.9  83.00   56.62702  38.17887  0.883  0.7630075
#> 19  19.52836200  0.42200      8.6  98.00   28.36795  17.35676  0.387  1.5624119
#> 20  20.54324117  0.41900     14.8  97.30  173.61590  39.11952  3.210  4.1655816
#> 21   0.00000200 22.30000      7.2   5.00  222.41860  54.06682 19.300  4.4459418
#> 22   1.33225167  0.75100      6.5   4.00  288.98060 294.28220 41.800 14.6034747
#> 23   3.39668933  2.17000     15.6  90.00  139.99370 338.09850  6.100  1.9605859
#> 24  49.32954433  4.22000      9.2  93.00  658.19810 202.53980 10.300 16.2027497
#> 25   0.00003600 34.50000     20.2  99.00  732.20780 351.42530 37.100 10.5489456
#> 26   3.08950333 12.70000       NA  65.50   55.04856  16.85995  4.990  2.5553953
#> 27   2.18452350  0.13500       NA  30.00   20.75014   5.79267  4.190  0.6288657
#> 28   0.00003600 12.50000     28.6  99.00  568.99200 200.84810 26.700  8.7110922
#> 29   2.04784200  0.00841       NA   5.00    8.95527   1.37855  1.750  0.1662127
#> 30   7.49402750  0.98000     14.6  91.00   42.12889  12.35732  0.965  1.1467629
#> 31   6.91123783  0.48200     11.6  96.00   30.97279 165.79100  7.990  2.8993036
#> 32   5.68129900  0.37700     16.6  90.00   28.29194   7.25574  0.444  1.0374761
#> 33   1.06719433  0.07190     12.8 100.00   13.27830  20.06847  0.179  0.2330044
#> 34   0.39678050 24.10000       NA   0.00   31.70398   6.24003  6.700  0.9065301
#> 35   1.52714167  0.02670       NA  90.00    7.81803   2.72198  2.100  0.3115324
#> 36   0.06733867 26.30000      8.9  71.00  278.03190  72.93042 21.900  7.3896926
#> 37  44.09012050 23.00000     17.4  99.00  770.39350 346.55110 13.900  4.8949538
#> 38  19.54137450 94.80000     23.5  99.00  142.60950  83.35802  4.570  2.2852030
#> 39   0.00000000  0.02260     14.7  88.00   48.20846  26.55097  1.850  2.3361618
#> 40   0.00301600  0.19500       NA  16.00   45.28752  11.00249  6.950  3.9638222
#> 41   0.00000000  0.68300      5.5  39.00  118.56180  55.21572  6.130  8.2693569
#> 42  26.34419950  3.68000     12.6 100.00  363.01320  82.74544 10.100  6.2945847
#> 43  10.12481567  0.57300     12.9  94.30  102.49150  43.77949  2.330  5.2309644
#> 44   8.26470100  0.37600     17.0  72.00  115.31850  31.05017  4.050  4.7341943
#> 45  14.18802800 19.80000     11.8  50.00  343.85040 122.55080 17.000  5.5610250
#> 46   0.00000200  9.43000     20.3 100.00  414.23890 304.71690 27.000  6.0574964
#> 47  18.81965283  4.63000     13.0  75.00  140.36600  16.24510  2.320  2.7675602
#> 48  21.65607267  0.44200     13.7  97.70   54.98977  11.79091  0.362  0.7759290
#> 49  42.47252133  1.69000     22.5  99.99  233.30670 132.10610  8.070  4.7452582
#> 50   2.25995317 24.50000     16.0  21.00  282.80140 108.01090 17.200  7.3177969
#> 51   3.52419900  0.77600      9.5   0.00  269.07660  30.53244 24.900  5.0849807
#>        ForPort Embs IGOs   UNVote CostImpEx Tariff TBTs TIRcon RTAs Visa  StMob
#> 1   2188.31100   82  196 38.46245     364.0   1.17  205      0   14   68 268.00
#> 2    808.68320   88  227 42.63920       0.0   1.60 1144      1   30   80  69.80
#> 3   1574.30800   84  248 43.00308       0.0   1.60 1348      1   30   80  42.30
#> 4      8.48896   52  145 38.60601     595.0  10.53   NA      0   12   45  50.40
#> 5     15.50880   67  209 42.95986      52.0   1.60 1140      1   30   79  28.80
#> 6           NA   41   85 38.50109     140.0   0.50    2      0   15   90   3.50
#> 7   2615.08900   82  222 42.64127     150.0   0.00  300      1   33   80  49.80
#> 8   1541.03500  100  193 37.13661     255.5   3.54 1230      1   17   11 445.00
#> 9     40.76228   43  172 42.34705     100.0   1.60 1141      1   30   79  29.40
#> 10   107.71030   84  201 42.22506       0.0   1.60 1456      1   30   79  47.90
#> 11  6386.91800  100  309 42.62615      45.0   1.60 1163      1   30   82 233.00
#> 12  1021.27800   77  259 42.79140       0.0   1.60 1393      1   30   81  33.20
#> 13  2040.57100   89  280 43.10103       0.0   1.60 1212      1   30   81  51.00
#> 14    17.41899   46  194 42.87997       0.0   1.60 1153      1   30   79   5.99
#> 15   747.62680   74  269 42.74629      70.0   1.60 1215      1   30   81  22.00
#> 16  6745.32800  100  329 40.42295       0.0   1.60 1385      1   30   81 152.00
#> 17  8206.97800  100  285 39.40118      25.0   1.60 1187      1   30   81 326.00
#> 18   185.10480   83  216 42.69301      30.0   1.60 1140      1   30   80  48.60
#> 19    16.88908   62  197 42.74966       0.0   1.60 1179      1   30   78   5.05
#> 20    72.52570   79  212 42.53351       0.0   1.60 1173      1   30   79  26.20
#> 21   287.61530   83  175 38.10477     303.2   2.64  118      1   17   59  31.10
#> 22   270.28240   95  198 36.92131     226.7   6.32  117      1   15   51 120.00
#> 23  6908.41200   74  195 42.68568     150.0   1.60 1143      1   30   80  24.50
#> 24  3264.08700   93  288 42.99491       0.0   1.60 1171      1   30   81  97.00
#> 25  7449.99300   95  201 42.17231     161.0   1.35  805      0   14   84 133.00
#> 26    87.49439   62  117 38.72687     320.0   2.51   19      1    2   48  83.90
#> 27          NA   42   95 38.02232     220.0   9.77    3      0   15   61   4.43
#> 28  1195.63500   86  210 42.00012      38.0   7.68  839      1   46   85  83.80
#> 29          NA   41   82 37.63291     350.0   1.65    1      0   16   61   6.46
#> 30    31.03612   50  190 42.71283      28.0   1.60 1170      1   30   79  12.10
#> 31 10601.75000   37  190 42.98426       0.0   1.60 1140      1   30   80  12.70
#> 32    26.82834   48  184 42.70066      35.0   1.60 1170      1   30   79   8.79
#> 33   139.86780   28  169 42.56485      25.0   1.60 1140      1   30   80   1.77
#> 34     0.04805   52   87 35.75572     350.0     NA    2      0   17   17   5.81
#> 35     5.20507   40  107 39.05257     147.0     NA    7      1    1   21   8.48
#> 36   248.16150   74  177 38.28493     105.0   3.68  236      0   17   91 115.00
#> 37  4697.46400   89  265 42.80321       0.0   1.60 1754      1   30   80  69.30
#> 38  1632.78100   78  259 43.11114       0.0   1.02   81      1   31   81  22.60
#> 39   235.98580   44  149 42.32222     147.0   1.27  108      0   13   80  49.20
#> 40    13.38957   72  157 38.14939     992.0   9.99  112      1   11    1  30.30
#> 41   103.68870   66  158 37.68160     103.0   3.40  256      0   17   57  10.60
#> 42   218.88300   87  238 42.58490       0.0   1.60 1147      1   30   79  35.60
#> 43   327.18220   72  236 43.18605       0.0   1.60 1141      1   30   80  14.40
#> 44    35.20750   75  217 42.52487       0.0   1.60 1230      1   30   79  38.70
#> 45   304.70170   99  232 36.12970     244.5   3.43   85      1    2   17 135.00
#> 46  1487.02000   58  119 38.74750      77.0   0.00   57      0   19   92  19.30
#> 47    70.29841   62  201 43.10103       0.0   1.60 1187      1   30   79  40.90
#> 48    49.05213   43  197 43.10103       0.0   1.60 1250      1   30   79   3.60
#> 49  1463.55900   78  278 42.75153      40.0   1.60 1364      1   30   81  28.10
#> 50   220.11210   78  152 38.82629     140.0     NA  609      0   17   61  30.30
#> 51     0.00189   80  138 37.53063     322.0   2.86  114      0   19   31  53.40
#>    Research    Pat CultServ CultGood Tourist MigStock        Lang       Renew
#> 1     46761  229.8  2.00098    6.807   8.263  5.66849 19.00543746  9.18050296
#> 2     14432  519.1  1.52758    8.932  28.121  1.30365 14.19041597 34.39499175
#> 3     20176  655.8  2.38787   12.698   7.481  1.32919 10.69085182  9.20164053
#> 4      2084     NA  0.01472       NA   0.125  5.08139  0.17744048 34.74706901
#> 5      1854   12.2  0.11991    0.880   8.252  0.69462  6.28997761 17.65007793
#> 6       229     NA       NA    0.364   0.219  0.14041  6.24168794  0.01493504
#> 7     30056 1315.7  1.15976   28.965   9.205  2.12634 13.63554873 25.29198368
#> 8     95919  927.8  2.88624   74.468  59.270  4.30982  0.73963574 12.41335257
#> 9      1337    7.4  0.13700    0.529   3.187  0.23838 12.32117982  9.94206602
#> 10     9190  112.6  0.32230    8.807   9.321  1.10637  7.72876772 14.82856195
#> 11    78353 2771.7  4.83603   47.161  35.555 10.92850 15.26134955 14.20625270
#> 12    15986  306.0  1.98816    4.912  10.781  0.54405 19.26996939 33.17027717
#> 13    39777  334.7       NA   11.759  75.315  3.34468  5.15846746 16.25409774
#> 14     1594   14.4  0.10275    0.745   3.147  0.32117 12.21152796 27.47683347
#> 15    10303  412.5  0.58878    1.858   2.789  0.46894 15.04353335 43.23526259
#> 16    56677 1469.0  7.01905   41.028  82.570  4.37339 11.50318182 13.49926232
#> 17    96337 1313.7  9.56996   35.139  35.814  8.91445 19.35213020  8.71168557
#> 18     8020   43.4  0.47483    1.626  24.799  1.06705 10.59013987 17.17004755
#> 19     2342   15.1  0.28331    0.737  13.809  0.54069 11.97641492 33.12746664
#> 20     5180  109.5  1.51071    2.614   5.302  0.84562  5.22493581 15.55936205
#> 21     2624   23.7  0.13405       NA  11.519  2.06821  0.03222059 36.87934810
#> 22    20234  437.6  3.29198    7.991  14.569  8.43759  3.77194295 36.02122257
#> 23     7451  140.6  0.52220    3.424  10.100  1.28606 19.46425530  9.08139901
#> 24    45825  513.0  0.57832   26.843  52.372  4.87039  6.40814649 16.51685058
#> 25    31184  583.5  2.20754   16.182  24.040  2.19471  0.03107851  6.29735708
#> 26     1026    3.6  0.05780    0.985   6.509  6.38381  7.20274577  1.55844291
#> 27      378     NA  0.00501       NA   5.012  0.93047  4.41049248 64.92377591
#> 28    20437  249.8  1.79800       NA  17.242  2.06535  2.13461064  2.70770280
#> 29      175     NA       NA       NA   3.315  1.12532  0.16638354 59.31586075
#> 30     1482    7.5  0.06532    1.281   2.296  0.57490 10.50042321 28.96117925
#> 31     1206   66.2  8.04500    0.636   1.054  0.29738 21.48526056  9.03131366
#> 32      794    1.5  0.03320    0.735   1.793  0.47182 12.00784247 38.09801717
#> 33      347    4.7  2.46272    0.369   1.966  0.11826 16.91195826  5.35500475
#> 34      299     NA  0.08905       NA   2.907  2.57501  0.01884848 61.52781165
#> 35      293    0.3  0.00266    0.046   0.404  0.08170  2.25185464  3.42973729
#> 36     8080   64.2  1.15292    7.555  26.757  4.00744  6.66895073  5.19444259
#> 37    33445  641.2  2.42317   19.159  15.828  1.68457 18.15321190  5.88946308
#> 38    11696  168.1  1.66163    2.981   5.960  0.66093 15.09626505 57.77200160
#> 39     7731   46.5  0.34615    1.213   3.370  1.48606 16.68013129 30.78917555
#> 40     7122    7.2  0.03553    1.256   0.966  4.09286  1.71805798 46.47632492
#> 41     1361   11.3  0.29555    3.185   5.967  1.40403 10.89425114 27.45154445
#> 42    13008  179.3  1.44714   11.912  17.471  4.24078  9.10344746 11.91148821
#> 43    10943   46.1  0.53946    2.616  11.223  1.75075  6.93615852 27.15729839
#> 44     4173   42.6  0.14490    2.387  10.223  3.32536  7.05215979 23.69779403
#> 45    16182  141.5  1.44633    8.379  24.571  7.65393  5.92039782  3.30422835
#> 46    11411  270.5  0.92780   14.507  12.914  2.47645 12.57967365  0.70863530
#> 47     2741   44.2  0.08820    2.407   2.027  0.49825  8.23424298 13.40916822
#> 48     2653   30.8  0.20211    1.212   3.032  0.18037 14.10243000 20.87739614
#> 49    23514  661.5  0.90896    6.014   6.782  1.03933 18.16742472 53.24776980
#> 50     5317   53.6  0.08969    6.661  32.530  4.13663  0.15905946 22.86307013
#> 51     3618     NA       NA       NA  10.013  1.16007  0.08500231 34.99856984
#>     PrimEner        CO2    MatCon     Forest Poverty Palma TertGrad FreePress
#> 1  123.01003 15.3701378 38.381073  6.3295294     0.3  1.40 29.78425        22
#> 2   84.92785  6.8737132 15.763616  3.0382810     0.7  1.11 13.08167        22
#> 3  113.76871  8.3281599 15.884942  5.7098277     0.0  0.96 31.70948        12
#> 4   74.74533  0.4591420  2.577148  5.2605370    19.6  1.26       NA        62
#> 5  151.98183  5.8716159 18.704688  1.5483901     1.5  1.49 24.47098        42
#> 6  113.44995 22.1247012 23.331756  3.2269754      NA    NA       NA        76
#> 7   53.47854  4.3115630 13.520983  0.9916749     0.0  1.20 36.94929        13
#> 8  175.31089  7.5439076 23.647067  3.7251454     1.4  2.10  3.57765        87
#> 9   77.69450  5.2603520 24.426326  1.7662068     0.0  1.40 25.54672        23
#> 10 134.44935  9.1659784 17.115225  4.9847983     0.0  0.92 19.34180        21
#> 11  86.76327  8.8893704 15.015616  2.3703489     0.0  1.10 24.74205        20
#> 12  63.75411  5.9357125 16.651923  5.3043935     0.2  1.04 30.38373        12
#> 13  79.00655  5.0338245 11.998735  6.9265991     1.0  1.46 29.29108        28
#> 14 169.38370 14.8488192 32.980617  9.4062132     0.5  1.26 37.60498        16
#> 15 159.24526  8.6607212 24.605785  8.5039377     0.0  0.95 22.50554        12
#> 16  97.46473  4.5720882 11.827461  3.7879808     0.0  1.31 17.66307        26
#> 17  72.59383  6.4974405  8.495755  5.4662193     0.2  1.25 29.33262        25
#> 18  88.19577  6.1803373 11.624611  1.8997660     1.5  1.55 22.92907        44
#> 19  94.25010  3.9738049  9.791906  1.7280976     0.7  1.22 18.25773        41
#> 20  95.79304  4.2655750 16.923169  5.8471809     0.5  1.13 20.67309        44
#> 21  88.36380  1.8193633  7.190287 10.6347954     7.2  1.81  8.48301        49
#> 22 118.26344  1.7300004  5.337525  3.3552538    21.2  1.48  9.13629        43
#> 23  56.06685  7.3781178 13.938004  3.5608942     0.5  1.25 26.83780        18
#> 24  71.12753  5.2708668 10.904800  2.0543587      NA  1.42 14.35266        31
#> 25  92.99299  9.5387061  9.381838  1.4232737      NA  1.22 29.87681        27
#> 26 188.00931 14.3623897 28.212457  0.4801006     0.1  0.92       NA        85
#> 27 133.44976  0.4377600  4.882918 22.8067646      NA  1.16       NA        70
#> 28 158.23984 11.5703454 15.904725  3.0383172     0.3    NA 24.35774        34
#> 29        NA  0.2972009 10.606696 13.1990072    22.7  1.68       NA        85
#> 30  90.93856  4.3780901 14.803791  8.0974972     0.7  1.43 33.09735        21
#> 31  73.25911 17.3621214 28.154633  4.5446296     0.2  1.40 34.27727        14
#> 32  97.98307  3.4981929 16.002621 10.3812689     0.7  1.44 27.79356        26
#> 33  55.99671  5.4915248 16.261072  2.9263320     0.0    NA 13.28486        23
#> 34  77.95974  0.4166004  3.300184  6.3602604     6.4    NA       NA        73
#> 35 161.97873  7.1273263 33.493893  2.9419658     0.2  1.23 23.71332        37
#> 36 122.65287  8.0329916 18.856906 17.2735413      NA  2.62 16.36954        69
#> 37  94.71582  9.9201381 14.304186  1.8284548     0.0  0.99 28.86907        11
#> 38  88.23266  9.2709451 21.150196  3.1175512     0.2  0.88 26.65009         8
#> 39 132.29055  7.6865758 24.861794  6.0694910      NA    NA 26.00903        19
#> 40 105.86127  0.8962641  4.432336  0.3263907     6.1  1.18  8.75331        65
#> 41  72.31282  1.0554568  3.972959  4.4737686     8.3  2.17 16.99506        44
#> 42 101.58949  7.5171516 17.682962  5.8058431     0.0  1.23 23.70020        34
#> 43  78.17923  4.3315540 11.090858 31.8455434     0.5  1.50 18.18253        17
#> 44  80.93064  3.5161537 11.469474  2.4348105     0.0  0.96 13.01408        38
#> 45 192.45083 11.8575278 16.362004  5.3007501     0.0  2.02  2.00915        83
#> 46  63.77904 10.3063319 33.369013  5.9836906      NA    NA 28.21053        67
#> 47 107.86727  5.6615813 10.910721  5.0249404     0.7  0.88 18.92215        26
#> 48 113.88117  6.2136868 13.770686  2.0597113     0.0  0.88 18.82353        23
#> 49 112.44515  4.4781822 17.027679  8.5016067     0.5  0.93 22.99885        11
#> 50 132.60433  4.6218600 12.304692  6.7611565     0.0  1.65 13.12545        77
#> 51 130.35543  1.8035657 13.594893 12.7836337     2.6  1.64       NA        84
#>      TolMin NGOs CPI    FemLab WomParl PubDebt  PrivDebt     GDPGrow   RDExp
#> 1  3.600000  311  77 0.8592661 28.6667   37.64 202.50016  0.34888036 2.20207
#> 2  4.800000  301  75 0.8872138 30.6011   86.19 141.28513  2.18677558 3.07181
#> 3  4.600000 1824  75 0.8699724 38.0000  106.05 225.72427  1.36616464 2.45668
#> 4  8.400000   16  28 0.5414029 20.2857   31.67  40.81502  6.16478747      NA
#> 5  5.100000   28  43 0.8927631 19.1667   26.30 116.98938  4.32022261 0.95657
#> 6  5.600000    1  62        NA  9.0909    2.81        NA  0.02851029      NA
#> 7  3.600000  741  85 0.8936418 32.5000   45.66 237.77122 -0.01982295 2.96635
#> 8  7.600000  117  41 0.8359555 23.7030   42.92 193.43419  6.30396712 2.06558
#> 9  6.000000   19  57 0.8633334 17.8571  108.87 345.59967  0.00000000 0.45585
#> 10 5.100000   55  57 0.8119483 20.0000   40.31  73.21357  4.04300124 1.94865
#> 11 4.900000  986  81 0.8844910 36.9841   70.99 105.91991  1.79471020 2.87749
#> 12 4.400000  208  88 0.9270000 37.4302   45.53 223.23741  1.50309097 3.01387
#> 13 5.800000  341  57 0.8631487 39.1429   99.26 165.79129  2.85707386 1.21961
#> 14 7.536316   21  71 0.9000751 26.7327    9.74 126.58231  4.87884628 1.49502
#> 15 1.500000  118  85 0.9651910 42.0000   62.51 181.23815  2.33629470 2.90474
#> 16 7.000000 1071  70 0.8953486 25.8232   96.14 188.68200  1.42667764 2.23135
#> 17 6.400000 1582  82 0.8697419 30.0000   88.98 169.14574  1.12929130 1.70304
#> 18 5.100000  102  48 0.7784910 18.3333  176.94 124.65858  1.49779228 0.95673
#> 19 5.500000   30  49 0.8664952 19.8675   86.67 130.34435  3.99021292 0.85439
#> 20 4.500000   62  45 0.8249007 10.0503   75.33  92.51395  4.33844641 1.37762
#> 21 7.200000   32  37 0.6149829 19.8214   27.30  40.83840  3.92307260 0.08466
#> 22 8.300000  167  40 0.3455764 11.8081   69.07  55.03717  5.42924312 0.62740
#> 23 1.100000   64  74 0.8078358 22.1519   78.68 299.06965  6.49733804 1.51368
#> 24 4.800000  498  50 0.7364464 30.9524  132.71 115.32313  1.62982757 1.33488
#> 25 3.400000  252  73 0.7796105  9.2632  247.98 158.42065  1.88001976 3.28363
#> 26 7.900000    5  31 0.9010363 27.1028   21.89  52.65279  2.59552272 0.16944
#> 27 6.600000    0  21 0.8788125 20.3252   32.54  68.15849  5.19299957 0.11827
#> 28 2.600000  145  54 0.7311271 17.0000   37.89 193.24990  2.62122194 4.22816
#> 29 6.406549    1  29 1.0274716 27.5168   61.86  19.86313  5.33713473      NA
#> 30 4.100000   16  59 0.9398145 21.2766   42.78  63.30250  5.31639215 1.04246
#> 31 2.800000   66  82 0.8341093 28.3333   21.45 421.38527 -0.67739873 1.28773
#> 32 8.300000   18  58 0.9234106 16.0000   34.94  96.90818  5.56194830 0.62550
#> 33 3.600000   14  56 0.6323963 12.5000   63.96 190.06811  4.14990811 0.76923
#> 34 9.800000    1  30        NA 10.1617   34.34  21.86631  5.40145516      NA
#> 35 3.500000    3  36 0.8351940 17.1053      NA  85.61135  4.22356684 0.15501
#> 36 6.200000  113  47 0.6499864 10.3604   57.39 145.18617  4.43836861 1.29813
#> 37 4.500000  731  82 0.8748585 38.0000   65.12 231.32965  2.54478724 2.01329
#> 38 3.400000  160  85 0.9480041 39.6450   27.94 252.12094  0.99831301 1.93301
#> 39 3.300000   54  89 0.8854174 34.1667   29.95 174.09862  0.86445133 1.15236
#> 40 9.700000   25  32 0.2970086 20.5882   63.57  19.67393  3.65519604 0.24554
#> 41 8.200000  104  34 0.6490328 29.4521   34.79  56.33445  5.05974739 0.13839
#> 42 5.900000   50  60 0.8201792 28.0435   51.28  86.22427  4.53454572 1.00337
#> 43 2.200000   81  63 0.9104053 34.7826  128.98 184.73320  2.99637372 1.27885
#> 44 6.500000   30  48 0.7703541 20.6687   39.30  28.38030  7.57933150 0.48765
#> 45 8.500000   82  29 0.8649159 15.7778   16.42  70.67039  0.00000000 1.13202
#> 46 2.300000  139  84 0.7977098 23.7624  104.68 165.48938  3.52678166 2.19754
#> 47 6.600000   24  50 0.8132671 20.0000   52.91  92.63818  3.22730973 1.17845
#> 48 4.500000   31  61 0.9055214 36.6667   83.15  88.14863  4.91413516 2.21174
#> 49 1.800000  299  84 0.9490853 43.5530   43.40 207.22266  0.81838277 3.26285
#> 50 8.200000   98  37 0.8148830  4.8583   42.72 130.82868  3.64110272 0.62720
#> 51 5.700000    8  35 0.9194402 26.7206   58.25 123.81487  5.72587785 0.37404
#>    NEET
#> 1   8.7
#> 2   6.5
#> 3   9.3
#> 4  27.4
#> 5  15.3
#> 6  17.2
#> 7   6.5
#> 8    NA
#> 9  16.1
#> 10  6.3
#> 11  6.3
#> 12  7.0
#> 13 13.3
#> 14  9.4
#> 15  9.4
#> 16 11.5
#> 17 10.3
#> 18 15.3
#> 19 15.4
#> 20 11.0
#> 21 21.5
#> 22 27.9
#> 23 10.9
#> 24 20.0
#> 25  3.5
#> 26  9.5
#> 27 12.7
#> 28   NA
#> 29 42.1
#> 30  9.2
#> 31  5.9
#> 32 10.3
#> 33  8.0
#> 34 17.4
#> 35 19.8
#> 36 12.8
#> 37  4.0
#> 38  4.6
#> 39 11.8
#> 40 30.4
#> 41 21.7
#> 42  9.5
#> 43  9.3
#> 44 15.2
#> 45 12.4
#> 46  4.0
#> 47 12.1
#> 48  6.5
#> 49  6.2
#> 50 15.0
#> 51  0.6